### 2023-12-11 #6137 @editorialbot
[PRE REVIEW]: ProMCDA: A Python package for Probabilistic Multi-Criteria Decision Analysis
**Submitting author:** @Flaminietta (<a href="http://orcid.org/0000-0003-0515-5282">Flaminia Catalli</a>)
**Repository:** https://github.com/wetransform-os/ProMCDA
**Branch with paper.md** (empty if default branch): main
**Version:** v1.0.8
