+++ README.md
@@ -4,14 +4,16 @@
 
 # Probabilistic Multi Criteria Decision Analysis
 
-<!-- [![status](https://joss.theoj.org/papers/4214c6e588774490458e34630e8052c1/status.svg)](https://joss.theoj.org/papers/4214c6e588774490458e34630e8052c1) -->
+
 <!-- [![PyPi version](https://img.shields.io/pypi/v/promcda?color=blue)](https://pypi.org/project/promcda) -->
 
 ![PyPI](https://img.shields.io/pypi/v/ProMCDA?label=pypi%20package)
-![PyPI - Downloads](https://img.shields.io/pypi/dm/ProMCDA)
-[![Downloads](https://static.pepy.tech/badge/ProMCDA)](https://pepy.tech/project/ProMCDA)
 [![pytest](https://github.com/wetransform-os/ProMCDA/actions/workflows/python-app-tests.yml/badge.svg)](https://github.com/wetransform-os/ProMCDA/actions/workflows/python-app-tests.yml)
 ![License](https://img.shields.io/badge/license-EPL%202.0-blue)
+[![status](https://joss.theoj.org/papers/cd66aa1ed9ff89b5519d977f4a16379d/status.svg)](https://joss.theoj.org/papers/cd66aa1ed9ff89b5519d977f4a16379d)
+![PyPI - Downloads](https://img.shields.io/pypi/dm/ProMCDA)
+[![Downloads](https://static.pepy.tech/badge/ProMCDA)](https://pepy.tech/project/ProMCDA)
+
 
 A tool to estimate scores of alternatives and their uncertainties based on a Multi Criteria Decision Analysis (MCDA) approach.
 
