### 2023-12-21 #6137 @Flaminietta
Hi @galessiorob! One suggestion for a second reviewer might be @paulrougieux who reviewed the CoinR JOSS paper cited in our references.
