### 2024-01-08 #6190 @editorialbot
[REVIEW]: ProMCDA: A Python package for Probabilistic Multi-Criteria Decision Analysis
**Submitting author:** @Flaminietta (<a href="http://orcid.org/0000-0003-0515-5282">Flaminia Catalli</a>)
**Repository:** https://github.com/wetransform-os/ProMCDA
**Branch with paper.md** (empty if default branch): main
**Version:** v1.1.0
