### 2024-01-09 wetransform-os/ProMCDA#18 @B3J4y
JOSS Paper: The Tables are Partially Visible
In my PDF, none of the tables are visible. @Flaminietta, could you please correct the source file so that the tables are viewable? Attached is a picture showing how the tables appear to me:

![half_visible_table](https://github.com/wetransform-os/ProMCDA/assets/1398738/63156be0-9a6c-46a9-aef8-0004c70fc11f)

Is this a problem specific to the PDF reader I'm using?
