### 2024-01-09 wetransform-os/ProMCDA#18 @Flaminietta
Hi @B3J4y, paper.pdf file (in the branch main of my repository) displays all tables just fine. The problem is in the preprint article and honestly, I am not quite sure what I can do to solve it. I can try to use different file extensions for the pictures and ask editorialbot to regenerate the preprint.
