### 2024-01-09 ffbb26b
Update paper.md
Use PDF files for tables in the paper.
