### 2024-01-09 6bfb49d
Update paper.md
Add tables in JPG format.
