### 2024-01-09 wetransform-os/ProMCDA#18 @Flaminietta
I've tried with PDF and JPG files but the problem in the preprint persists.
