### 2024-01-09 5af665c
Update paper.md
Try to use the latex command \begin{center}  and \end{center} to center the tables in the pdf file.
