### 2024-01-09 9e86abe
Update paper.md
Use \includegraphics[]{Table.png}.
