### 2024-01-09 4a74f5d
Update paper.md
Control size of pictures.
