### 2024-01-09 wetransform-os/ProMCDA#18 @Flaminietta
Hello again @B3J4y, the problem seems to be solved now: https://github.com/openjournals/joss-papers/actions/runs/7460439873.
