### 2024-01-09 wetransform-os/ProMCDA#18 @Flaminietta
> Thank you. How do I see the resulting PDF, now? 🙂

I can see it in the artifacts, either [in my workflow artifact](https://github.com/wetransform-os/ProMCDA/actions/runs/7460417330) or [in the preprint](https://github.com/openjournals/joss-papers/actions/runs/7460439873) created by the @editorialbot (but then you have to compile the tex file).
