### 2024-01-24 a4caebe
add chages suggested by Pycharm after inspection
