+++ tests/unit_tests/test_normalization.py
@@ -6,18 +6,20 @@ from numpy.testing import assert_almost_equal
 from mcda.utils import *
 from mcda.utility_functions.normalization import Normalization
 
+
 class TestNormalization(unittest.TestCase):
 
     @staticmethod
     def get_input_matrix():
         input_matrix = read_matrix("tests/resources/input_matrix_without_uncert.csv")
-        input_matrix_no_alternatives = input_matrix.drop(input_matrix.columns[0],axis=1) # drop first column with alternatives
-                                                                                         # in the code this is happening in mcda_run
+        input_matrix_no_alternatives = input_matrix.drop(input_matrix.columns[0],
+                                                         axis=1)  # drop first column with alternatives
+        # in the code this is happening in mcda_run
         return input_matrix_no_alternatives
 
     @staticmethod
     def get_input_polarities():
-        polarities = ("-","-","+","+","+","+") # same as in ProMCDA/configuration_without_robustness.json
+        polarities = ("-", "-", "+", "+", "+", "+")  # same as in ProMCDA/configuration_without_robustness.json
 
         return polarities
 
@@ -31,9 +33,9 @@ class TestNormalization(unittest.TestCase):
         expected_res_no0 = read_matrix('tests/resources/normalization/res_minmax_no0.csv')
         norm = Normalization(input_matrix_no_alternatives, polarities)
         res_01 = norm.minmax(feature_range=(0, 1))
-        res_01.columns = res_01.columns.astype('str') # to match the type of columns in the two dfs
+        res_01.columns = res_01.columns.astype('str')  # to match the type of columns in the two dfs
         res_no0 = norm.minmax(feature_range=(0.1, 1))
-        res_no0.columns = res_no0.columns.astype('str') # to match the type of columns in the two dfs
+        res_no0.columns = res_no0.columns.astype('str')  # to match the type of columns in the two dfs
 
         # Then
         assert isinstance(res_01, pd.DataFrame)
@@ -41,10 +43,9 @@ class TestNormalization(unittest.TestCase):
         assert_frame_equal(res_01, expected_res_01, check_like=True, check_dtype=False)
         assert_frame_equal(res_no0, expected_res_no0, check_like=True, check_dtype=False)
         assert_almost_equal(res_01.to_numpy().min(), 0)
-        assert_almost_equal(res_01.to_numpy().max(),1)
-        assert_almost_equal(res_no0.to_numpy().min(),0.1)
-        assert_almost_equal(res_no0.to_numpy().max(),1)
-
+        assert_almost_equal(res_01.to_numpy().max(), 1)
+        assert_almost_equal(res_no0.to_numpy().min(), 0.1)
+        assert_almost_equal(res_no0.to_numpy().max(), 1)
 
     def test_target(self):
         # Given
@@ -54,11 +55,11 @@ class TestNormalization(unittest.TestCase):
         # When
         expected_res_01 = read_matrix('tests/resources/normalization/res_target_01.csv')
         expected_res_no0 = read_matrix('tests/resources/normalization/res_target_no0.csv')
-        norm = Normalization(input_matrix_no_alternatives,polarities)
+        norm = Normalization(input_matrix_no_alternatives, polarities)
         res_01 = norm.target(feature_range=(0, 1))
-        res_01.columns = res_01.columns.astype('str') # to match the type of columns in the two dfs
+        res_01.columns = res_01.columns.astype('str')  # to match the type of columns in the two dfs
         res_no0 = norm.target(feature_range=(0.1, 1))
-        res_no0.columns = res_no0.columns.astype('str') # to match the type of columns in the two dfs
+        res_no0.columns = res_no0.columns.astype('str')  # to match the type of columns in the two dfs
 
[…]
