### 2024-01-25 afcd069
address style violations raised by pylint
