+++ mcda/utils/utils_for_main.py
@@ -1,24 +1,34 @@
 import argparse
-import io
 import json
 import logging
 import os
 import pickle
 import random
+import sys
 from datetime import datetime
 from typing import List
 
 import numpy as np
 import pandas as pd
-import plotly.graph_objects as go
-import plotly.io as pio
-from PIL import Image
 from sklearn import preprocessing
 from sklearn.preprocessing import MinMaxScaler
 
-# logging.basicConfig(level=logging.WARNING)
-# suppress the debug messages produced by PIL internal logging
 logging.getLogger('PIL').setLevel(logging.WARNING)
+
+FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
+logger = logging.getLogger("ProMCDA")
+
+
+def check_config_error(condition, error_message):
+    if condition:
+        logger.error('Error Message', stack_info=True)
+        raise ValueError(error_message)
+
+
+def check_config_setting(condition, information_message):
+    if condition:
+        logger.info(information_message)
 
 
 def read_matrix(input_matrix_path: str) -> pd.DataFrame():
@@ -72,13 +82,13 @@ def save_df(df: pd.DataFrame, folder_path: str, filename: str):
     df.to_csv(path_or_buf=result_path, index=False)
 
 
-def save_dict(dict: dict, folder_path: str, filename: str):
+def save_dict(dictionary: dict, folder_path: str, filename: str):
     timestamp = datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
     new_filename = f"{timestamp}_{filename}"
     result_path = os.path.join(folder_path, new_filename)
 
     with open(result_path, 'wb') as fp:
-        pickle.dump(dict, fp)
+        pickle.dump(dictionary, fp)
 
 
 def save_config(config: dict, folder_path: str, filename: str):
@@ -211,164 +221,3 @@ def check_if_pdf_is_uniform(marginal_pdf: list) -> list:
     uniform_pdf_mask = [1 if pdf == 'uniform' else 0 for pdf in marginal_pdf]
 
     return uniform_pdf_mask
-
-
-def plot_norm_scores_without_uncert(scores: pd.DataFrame) -> object:
-    num_of_combinations = scores.shape[1] - 1
-    fig = go.Figure(layout_yaxis_range=[scores.iloc[:, 1:].values.min() - 0.5, scores.iloc[:, 1:].values.max() + 0.5],
-                    layout_yaxis_title="MCDA normalized score")
-    i = 0
-    while i <= num_of_combinations - 1:
-        fig.add_trace(go.Bar(
-            name=scores.columns[i + 1],
-            x=scores['Alternatives'].values.tolist(),
-            y=scores.iloc[:, i + 1],
-            showlegend=True
-        ))
-        i = i + 1
-    fig.update_traces(showlegend=True)
-    fig.update_layout(barmode='group', height=600, width=1000,
-                      title='<b>MCDA analysis<b>',
-                      title_font_size=22,
-
-                      xaxis=dict(
-                          tickmode="array",
-                          tickvals=np.arange(
-                              0, len(scores['Alternatives'][:])),
-                          ticktext=scores['Alternatives'][:],
-                          tickangle=45),
-                      yaxis=dict(
-                          range=[scores.iloc[:, 1:].values.min() - 0.5, scores.iloc[:, 1:].values.max() + 0.5])
-                      )
-    fig.show()
-    return fig
-
-
-def plot_non_norm_scores_without_uncert(scores: pd.DataFrame) -> object:
[…]
