+++ mcda/utils/utils_for_plotting.py
@@ -1,0 +1,170 @@
+import io
+import os
+from datetime import datetime
+
+import numpy as np
+import pandas as pd
+import plotly.graph_objects as go
+import plotly.io as pio
+from PIL import Image
+
+
+def plot_norm_scores_without_uncert(scores: pd.DataFrame) -> object:
+    num_of_combinations = scores.shape[1] - 1
+    fig = go.Figure(layout_yaxis_range=[scores.iloc[:, 1:].values.min() - 0.5, scores.iloc[:, 1:].values.max() + 0.5],
+                    layout_yaxis_title="MCDA normalized score")
+    i = 0
+    while i <= num_of_combinations - 1:
+        fig.add_trace(go.Bar(
+            name=scores.columns[i + 1],
+            x=scores['Alternatives'].values.tolist(),
+            y=scores.iloc[:, i + 1],
+            showlegend=True
+        ))
+        i = i + 1
+    fig.update_traces(showlegend=True)
+    fig.update_layout(barmode='group', height=600, width=1000,
+                      title='<b>MCDA analysis<b>',
+                      title_font_size=22,
+
+                      xaxis=dict(
+                          tickmode="array",
+                          tickvals=np.arange(
+                              0, len(scores['Alternatives'][:])),
+                          ticktext=scores['Alternatives'][:],
+                          tickangle=45),
+                      yaxis=dict(
+                          range=[scores.iloc[:, 1:].values.min() - 0.5, scores.iloc[:, 1:].values.max() + 0.5])
+                      )
+    fig.show()
+    return fig
+
+
+def plot_non_norm_scores_without_uncert(scores: pd.DataFrame) -> object:
+    num_of_combinations = scores.shape[1] - 1
+    fig = go.Figure(layout_yaxis_title="MCDA rough score")
+    i = 0
+    while i <= num_of_combinations - 1:
+        fig.add_trace(go.Bar(
+            name=scores.columns[i + 1],
+            x=scores['Alternatives'].values.tolist(),
+            y=scores.iloc[:, i + 1],
+            showlegend=True
+        ))
+        i = i + 1
+    fig.update_traces(showlegend=True)
+    fig.update_layout(barmode='group', height=600, width=1000,
+                      title='<b>MCDA analysis<b>',
+                      title_font_size=22,
+
+                      xaxis=dict(
+                          tickmode="array",
+                          tickvals=np.arange(
+                              0, len(scores['Alternatives'][:])),
+                          ticktext=scores['Alternatives'][:],
+                          tickangle=45)
+                      )
+    fig.show()
+    return fig
+
+
+def plot_mean_scores(all_means: pd.DataFrame, all_stds: pd.DataFrame, plot_std: str, rand_on: str) -> object:
+    num_of_combinations = all_means.shape[1] - 1
+    fig = go.Figure(layout_yaxis_title="MCDA average scores and std")
+    i = 0
+    while i <= num_of_combinations - 1:
+        if plot_std == "plot_std":
+            fig.add_trace(go.Bar(
+                name=all_means.columns[i + 1],
+                x=all_means['Alternatives'].values.tolist(),
+                y=all_means.iloc[:, i + 1],
+                error_y=dict(type='data', array=all_stds.iloc[:, i + 1])
+            ))
+            fig.update_layout(title=f'<b>MCDA analysis with added randomness on the {rand_on}<b> (rough scores)',
+                              title_font_size=22)
+        else:
+            fig.add_trace(go.Bar(
+                name=all_means.columns[i + 1],
+                x=all_means['Alternatives'].values.tolist(),
+                y=all_means.iloc[:, i + 1],
+                showlegend=True
+            ))
[…]
