### 2024-01-31 wetransform-os/ProMCDA#23 @paulrougieux
Paper: add plots to illustrate some of the package's functionality 
Comments on JOSS_paper/paper.md in the version available at commit
11a34a0 “add paper preprint” by Flaminia Catalli.

> decision makers (DMs) Acronym is only used two other times. Just keep it as a word
> "decision makers" it's less confusing.

Can you illustrate the paper with a couple of plots? See for example the plots done in
the [JOSS paper describing the Coinr package](https://joss.theoj.org/papers/10.21105/joss.04567).

There are several plotting functions in mcda/utils.py

    plot_norm_scores_without_uncert
    plot_non_norm_scores_without_uncert
    plot_mean_scores
    plot_mean_scores_iterative

These functions could be used with sample data to generate a plot illustrating some of
the functionality of the package.

This is related to https://github.com/openjournals/joss-reviews/issues/6190
