### 2024-01-31 wetransform-os/ProMCDA#24 @paulrougieux
Document function arguments
- Function arguments should be documented whenever possible. The functions in these scripts are not documented:

    - [mcda/utils.py](https://github.com/wetransform-os/ProMCDA/blob/main/mcda/utils.py)
    - [mcda/mcda_with_robustness.py](https://github.com/wetransform-os/ProMCDA/blob/main/mcda/mcda_with_robustness.py) please document function arguments and provide examples of how to use them.

- Use a tool such as [pdoc](https://pdoc.dev/) or [sphynx](https://www.sphinx-doc.org/en/master/) to generate the documentation automatically from function docstrings. You can create a github workflow that generates the documentation and places it on a github page.

- Functions should be accompanied with an example of how to use them, based on a sample  data set. If several function depend on each other, it is fine to group those examples  under the main relevant function or object.

Related to the JOSS review https://github.com/openjournals/joss-reviews/issues/6190 .
