### 2024-02-01 wetransform-os/ProMCDA#25 @paulrougieux
Include sample data inside the package
Sample data should be stored within the package so that users can run code examples.

See https://stackoverflow.com/questions/7522250/how-to-include-package-data-with-setuptools-distutils

Related to issue https://github.com/wetransform-os/ProMCDA/issues/19 and to the JOSS review https://github.com/openjournals/joss-reviews/issues/6190
