### 2024-02-02 wetransform-os/ProMCDA#26 @B3J4y
Add Contribution Guidelines
To enable active participation with this package, it's essential to establish contribution guidelines.

A good example of how to effectively implement this can be found at: https://github.com/citation-file-format/citation-file-format. This repository includes clear statements in the README.md, alongside a CONTRIBUTING.md and a Code_OF_CONDUCT.md. While it's not necessary to be overly verbose, providing some guidance on the expected interactions would be beneficial.
