### 2024-02-02 wetransform-os/ProMCDA#27 @B3J4y
Cannot terminate the process
After running the toy example, the process did not finish, and I was unable to terminate it properly.
