### 2024-02-02 1c36549
continue refactoring of mcda_run and add doctrings
