+++ mcda/utils/utils_for_main.py
@@ -20,7 +20,21 @@ logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
 logger = logging.getLogger("ProMCDA")
 
 
-def check_config_error(condition, error_message):
+def check_config_error(condition: bool, error_message: str):
+    """
+    Check a condition and raise a ValueError with a specified error message if the condition is True.
+
+    Parameters:
+    - condition (bool): The condition to check.
+    - error_message (str): The error message to raise if the condition is True.
+
+    Raises:
+    - ValueError: If the condition is True, with the specified error message.
+
+    :return: None
+    :param condition: bool
+    """
+
     if condition:
         logger.error('Error Message', stack_info=True)
         raise ValueError(error_message)
@@ -31,7 +45,199 @@ def check_config_setting(condition, information_message):
         logger.info(information_message)
 
 
+def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: pd.DataFrame,
+                                   is_robustness_indicators: int, is_robustness_weights: int, polar: List[str],
+                                   mc_runs: int, num_indicators: int):
+    """
+    Process indicators and weights based on input parameters in the configuration.
+
+    Parameters:
+    - input_matrix_no_alternatives (DataFrame): The input matrix.
+    - is_robustness_indicators (int): Flag indicating whether the matrix should include indicator uncertainties
+      (0 or 1).
+    - is_robustness_weights (int): Flag indicating whether robustness analysis is considered for the weights
+      (0 or 1).
+    - marginal_pdf (list): List of marginal probability density functions for indicators.
+    - mc_runs (int): Number of Monte Carlo runs for robustness analysis.
+    - num_indicators: the number of indicators in the input matrix.
+    - config: the configuration dictionary.
+
+    Raises:
+    - ValueError: If there are duplicated rows in the input matrix or if there is an issue with the configuration.
+
+    Notes:
+    - For is_robustness_indicators == 0:
+    - Identifies and removes columns with constant values.
+    - Logs the number of alternatives and indicators.
+
+    - For is_robustness_indicators == 1:
+    - Handles uncertainty in indicators.
+    - Logs the number of alternatives and indicators.
+
+    - For is_robustness_weights == 0:
+    - Processes fixed weights if given.
+    - Logs weights and normalized weights.
+
+    - For is_robustness_weights == 1:
+    - Performs robustness analysis on weights.
+    - Logs randomly sampled weights.
+
+    :return: None
+    :param mc_runs: int
+    :param polar: List[str]
+    :param is_robustness_weights: int
+    :param is_robustness_indicators: int
+    :param input_matrix_no_alternatives: pd.DataFrame
+    :param config: dict
+    :param num_indicators: int
+    """
+
+    num_unique = input_matrix_no_alternatives.nunique()
+    cols_to_drop = num_unique[num_unique == 1].index
+    col_to_drop_indexes = input_matrix_no_alternatives.columns.get_indexer(cols_to_drop)
+
+    if is_robustness_indicators == 0:
+        _handle_no_robustness_indicators(input_matrix_no_alternatives)
+    else:  # matrix with uncertainty on indicators
+        logger.info("Number of alternatives: {}".format(input_matrix_no_alternatives.shape[0]))
+        logger.info("Number of indicators: {}".format(num_indicators))
+        # TODO: eliminate indicators with constant values (i.e. same mean and 0 std) - optional
+
[…]
