### 2024-02-14 853c1ce
complete refactoring of mcda_run
