+++ mcda/mcda_run.py
@@ -22,32 +22,23 @@ FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
 logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
 logger = logging.getLogger("ProMCDA")
 
-
-class UserStoppedInfo(Exception):
-    """A class representing information about user stopping conditions"""
-    pass
-
-
 # noinspection PyTypeChecker
-def main(input_config: dict, user_input_callback=input):
+def main(input_config: dict):
     """
         Execute the ProMCDA (Probabilistic Multi-Criteria Decision Analysis) process.
 
         Parameters:
         - input_config (dictionary): Configuration parameters for the ProMCDA process.
-        - user_input_callback (function, optional): Callback function for user input.
 
         Raises:
         - ValueError: If there are issues with the input matrix, weights, or indicators.
-        - UserStoppedInfo: If the user chooses to stop the process during execution.
 
         This function performs the ProMCDA process based on the provided configuration.
-        It handles various aspects such as sensitivity analysis, robustness analysis,
-        and uncertainty in indicators. The results are saved in output files, and plots
-        are generated to visualize the scores and rankings.
+        It handles various aspects such as the sensitivity analysis and the robustness analysis.
+        The results are saved in output files, and plots are generated to visualize the scores and rankings.
 
-        Note: Ensure that the input matrix, weights, and indicators are correctly specified
-        in the input configuration.
+        Note: Ensure that the input matrix, weights, polarities and indicators (with or without uncertainty)
+        are correctly specified in the input configuration.
         """
 
     is_sensitivity = None
@@ -56,10 +47,6 @@ def main(input_config: dict, user_input_callback=input):
     is_robustness_weights = 0
     marginal_pdf = []
     num_unique = []
-    iterative_random_w_means = {}
-    iterative_random_w_stds = {}
-    iterative_random_w_means_normalized = {}
-    iterative_random_w_stds_normalized = {}
 
     # Extracting relevant configuration values
     config = Config(input_config)
@@ -137,8 +124,8 @@ def main(input_config: dict, user_input_callback=input):
         num_indicators = (input_matrix_no_alternatives.shape[1] - num_non_exact_and_non_poisson)
 
     # Process indicators and weights based on input parameters in the configuration
-    process_indicators_and_weights(config, input_matrix_no_alternatives, is_robustness_indicators, is_robustness_weights,
-                                   polar, mc_runs, num_indicators)
+    polar, weights = process_indicators_and_weights(config, input_matrix_no_alternatives, is_robustness_indicators,
+                                                    is_robustness_weights, polar, mc_runs, num_indicators)
 
     # Check the number of indicators, weights, and polarities
     try:
@@ -147,300 +134,18 @@ def main(input_config: dict, user_input_callback=input):
         logging.error(str(e), stack_info=True)
         raise
 
-    # ----------------------------
-    # NO UNCERTAINTY OF INDICATORS   QUI, refactor!
-    # ----------------------------
+    # If there is no uncertainty of the indicators:
     if is_robustness_indicators == 0:
-        logger.info("Start ProMCDA without robustness of the indicators")
-        t = time.time()
-        scores = pd.DataFrame()
-        all_weights_means = pd.DataFrame()
-        mcda_no_uncert = MCDAWithoutRobustness(
-            config, input_matrix_no_alternatives)
-        # normalize the indicators
-        if is_sensitivity == "yes":
-            normalized_indicators = mcda_no_uncert.normalize_indicators()
-        else:
-            normalized_indicators = mcda_no_uncert.normalize_indicators(f_norm)
[…]
