+++ mcda/utils/utils_for_main.py
@@ -1,17 +1,20 @@
 import argparse
 import json
-import logging
-import os
 import pickle
 import random
-import sys
-from datetime import datetime
-from typing import List
+import time
+from typing import Union, Any, List
+from typing import Optional
 
-import numpy as np
 import pandas as pd
 from sklearn import preprocessing
 from sklearn.preprocessing import MinMaxScaler
+
+from mcda.configuration.config import Config
+from mcda.utils.utils_for_parallelization import *
+from mcda.utils.utils_for_plotting import *
+from mcda.mcda_without_robustness import MCDAWithoutRobustness
+from mcda.mcda_with_robustness import MCDAWithRobustness
 
 logging.getLogger('PIL').setLevel(logging.WARNING)
 
@@ -32,6 +35,7 @@ def check_config_error(condition: bool, error_message: str):
     - ValueError: If the condition is True, with the specified error message.
 
     :return: None
+    :param error_message: str
     :param condition: bool
     """
 
@@ -52,7 +56,7 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
     Process indicators and weights based on input parameters in the configuration.
 
     Parameters:
-    - input_matrix_no_alternatives (DataFrame): The input matrix.
+    - input_matrix (DataFrame): The input matrix.
     - is_robustness_indicators (int): Flag indicating whether the matrix should include indicator uncertainties
       (0 or 1).
     - is_robustness_weights (int): Flag indicating whether robustness analysis is considered for the weights
@@ -65,6 +69,11 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
     Raises:
     - ValueError: If there are duplicated rows in the input matrix or if there is an issue with the configuration.
 
+    Returns:
+    - a shorter list of polarities if one has been dropped together with the relative indicator,
+      which brings no information. Otherwise, the same list.
+    - the normalised weights (either fixed or random sampled weights, depending on the seetings)
+
     Notes:
     - For is_robustness_indicators == 0:
     - Identifies and removes columns with constant values.
@@ -76,13 +85,14 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
 
     - For is_robustness_weights == 0:
     - Processes fixed weights if given.
-    - Logs weights and normalized weights.
+    - Logs weights and normalised weights.
 
     - For is_robustness_weights == 1:
     - Performs robustness analysis on weights.
     - Logs randomly sampled weights.
 
-    :return: None
+    :return: polar, norm_weights
+    :rtype: List[str], Union[list, dict]
     :param mc_runs: int
     :param polar: List[str]
     :param is_robustness_weights: int
@@ -103,13 +113,20 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
         logger.info("Number of indicators: {}".format(num_indicators))
         # TODO: eliminate indicators with constant values (i.e. same mean and 0 std) - optional
 
-    _handle_polarities_and_weights(is_robustness_indicators, is_robustness_weights, num_unique, col_to_drop_indexes,
-                                   polar, config, mc_runs, num_indicators)
+    polar, norm_weights = _handle_polarities_and_weights(is_robustness_indicators, is_robustness_weights, num_unique,
+                                                         col_to_drop_indexes, polar, config, mc_runs, num_indicators)
+
+    return polar, norm_weights
 
 
 def _handle_polarities_and_weights(is_robustness_indicators: int, is_robustness_weights: int, num_unique,
                                    col_to_drop_indexes: np.ndarray, polar: List[str], config: dict, mc_runs: int,
[…]
