### 2024-02-15 wetransform-os/ProMCDA#28 @limengbinggz
Add an option to specify random seeds
To ensure reproducibility of results run by the functions with randomness in the package, could you add an option that allows users to set random seeds in the sampling process?

Related to the JOSS review [openjournals/joss-reviews#6190](https://github.com/openjournals/joss-reviews/issues/6190)
