### 2024-02-16 wetransform-os/ProMCDA#29 @limengbinggz
Minor issues in the paper
Typos:
- On page 2: the first bullet point - "normalisation" -> "normalization" (or just be consistent throughout all documents)
- Table 1: the formula of Min-max - $x_{ic}$ -> $x_{ia}$
- References: Becker et al. (2022) - The title shall be "composite indicators" instead of "compositeindicators".

Relevant literature:
- In the Statement of Need, how does the ProMCDA package compare to other Python libraries for MCDA, such as pymcdm and pyDecision?

Related to the JOSS review [openjournals/joss-reviews#6190](https://github.com/openjournals/joss-reviews/issues/6190)
