### 2024-02-16 wetransform-os/ProMCDA#30 @limengbinggz
Issues in README
- When running the toy example using the configuration.json file included in the package and specifying ``on_all_weights=yes``, the message displayed when running the code says "All weights are randomly sampled from a uniform distribution", although the README file says "all weights are simultaneously sampled from a normal distribution". Is this a typo?
- For arguments in the configuration file, could you give instructions on "given_weights"?
- In the README file, the list of strings for marginal distributions says "norm" for normal distribution, but in configuration.json, the string is "normal".

Related to the JOSS review [openjournals/joss-reviews#6190](https://github.com/openjournals/joss-reviews/issues/6190)
