+++ tests/unit_tests/test_utils.py
@@ -1,183 +1,0 @@
-import pandas as pd
-import unittest
-from unittest import TestCase
-
-from mcda.utils.utils_for_main import *
-from mcda.configuration.config import Config
-from mcda.utils.utils_for_main import _check_and_rescale_negative_indicators
-
-
-class TestUtils(unittest.TestCase):
-
-    @staticmethod
-    def get_test_config():
-        return {
-            "input_matrix_path": "/path/to/input_matrix.csv",
-            "polarity_for_each_indicator": ["-", "-", "+", "+"],
-            "sensitivity": {
-                "sensitivity_on": "yes",
-                "normalization": "minmax",
-                "aggregation": "weighted_sum"},
-            "robustness": {
-                "robustness_on": "yes",
-                "on_single_weights": "no",
-                "on_all_weights": "no",
-                "given_weights": [0.5, 0.5, 0.5, 0.5],
-                "on_indicators": "yes"},
-            "monte_carlo_sampling": {
-                "monte_carlo_runs": 10000,
-                "num_cores": 1,
-                "marginal_distribution_for_each_indicator": ['exact', 'uniform', 'normal', 'poisson']},
-            "output_path": "/path/to/output"
-        }
-
-    @staticmethod
-    def get_input_matrix_1() -> pd.DataFrame:
-        data = {'ind1': [1, 2, 3], 'ind2_min': [1, 2, 3], 'ind2_max': [5, 6, 7],
-                'ind3': [9, 10, 11], 'std3': [0.1, 0.1, 12], 'ind4_rate': [9, 10, 11]}
-        df = pd.DataFrame(data=data)
-
-        return df
-
-    @staticmethod
-    def get_input_matrix_2() -> pd.DataFrame:
-        data = {'ind1': [1, 2, 3], 'ind2_min': [-5, -6, -7], 'ind2_max': [5, 6, 7],
-                'ind3': [9, 10, 11], 'std3': [0.1, 0.1, 11], 'ind4_rate': [9, 10, 11]}
-        df = pd.DataFrame(data=data)
-
-        return df
-
-    @staticmethod
-    def get_input_matrix_3() -> pd.DataFrame:
-        data = {'ind1': [1, 2, 3], 'ind2_min': [6, -6, -7], 'ind2_max': [5, 6, 7],
-                'ind3': [9, 10, 11], 'std3': [0.1, 0.1, 11], 'ind4_rate': [9, 10, 11]}
-        df = pd.DataFrame(data=data)
-
-        return df
-
-    @staticmethod
-    def get_list_pdf() -> list:
-        list_pdf = ['exact', 'lnorm', 'norm', 'poisson', 'exact']
-        return list_pdf
-
-    @staticmethod
-    def get_input_matrix_negative() -> pd.DataFrame:
-        data = {'ind1': [1, -2, 3], 'ind2': [-5, -6, -7], 'ind3': [9, 10, -11], 'ind4': [9, 10, -11]}
-        df = pd.DataFrame(data=data)
-
-        return df
-
-    def test_randomly_sample_ix_weight(self):
-        # Given
-        num_weights = 4
-        index = 1
-        num_runs = 10
-
-        # When
-        out = randomly_sample_ix_weight(num_weights, index, num_runs)
-
-        # Then
-        isinstance(out, list)
-        assert len(out) == num_runs
-        assert [len(out[i]) == num_weights for i in range(num_weights)]
-        for j in range(num_runs):
-            for i in range(num_weights):
-                if i != index:
-                    assert out[j][i] == 1
-                else:
-                    assert out[j][i] != 1
-
-    def test_randomly_sample_all_weights(self):
-        # Given
-        num_weights = 4
-        num_runs = 10
-
-        # When
-        out = randomly_sample_all_weights(num_weights, num_runs)
-
[…]
