+++ mcda/utils/utils_for_main.py
@@ -84,7 +84,7 @@ def check_config_setting(condition: bool, information_message: str,
 
 def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: pd.DataFrame,
                                    is_robustness_indicators: int, is_robustness_weights: int, polar: List[str],
-                                   mc_runs: int, num_indicators: int) -> Tuple[List[str], Union[List[list], dict]]:
+                                   mc_runs: int, num_indicators: int) -> Tuple[List[str], Union[list, List[list], dict]]:
     """
     Process indicators and weights based on input parameters in the configuration.
 
@@ -157,7 +157,7 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
 def _handle_polarities_and_weights(is_robustness_indicators: int, is_robustness_weights: int, num_unique,
                                    col_to_drop_indexes: np.ndarray, polar: List[str], config: dict, mc_runs: int,
                                    num_indicators: int) \
-        -> Union[Tuple[List[str], List[str], None, None],
+        -> Union[Tuple[List[str], list, None, None],
         Tuple[List[str], None, List[List], None],
         Tuple[List[str], None, None, dict]]:
     """
@@ -191,14 +191,14 @@ def _handle_polarities_and_weights(is_robustness_indicators: int, is_robustness_
             return polar, None, norm_random_weights, None
         else:
             return polar, None, None, rand_weight_per_indicator
-        #  Return None for norm_fixed_weights
+        #  Return None for norm_fixed_weights and one of the other two cases of randomness
 
 
 def _handle_robustness_weights(config: dict, mc_runs: int, num_indicators: int) \
         -> Tuple[Union[List[list], None], Union[dict, None]]:
     """
     Handle the generation and normalization of random weights based on the specified settings
-    when a robustness analysis is requested on the weights.
+    when a robustness analysis is requested on all the weights.
     """
     norm_random_weights = []
     rand_weight_per_indicator = {}
@@ -796,7 +796,7 @@ def check_if_pdf_is_uniform(marginal_pdf: list) -> list:
 
 def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name: str, index_column_values: list,
                                            input_matrix: pd.DataFrame,
-                                           weights: Union[List[str], List[pd.DataFrame], dict, None],
+                                           weights: Union[list, List[list], dict],
                                            f_norm: str, f_agg: str):
     """
     Runs ProMCDA without uncertainty on the indicators, i.e. without performing
@@ -852,8 +852,9 @@ def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name
         # ALL RANDOMLY SAMPLED WEIGHTS (MCDA runs num_samples times)
         all_weights_score_means, all_weights_score_stds, \
             all_weights_score_means_normalized, all_weights_score_stds_normalized = \
-            _compute_scores_for_all_random_weights(normalized_indicators, is_sensitivity, index_column_name,
-                                                   index_column_values, input_matrix, weights, f_agg)
+            _compute_scores_for_all_random_weights(normalized_indicators, is_sensitivity,
+                                                   index_column_name, index_column_values,
+                                                   weights, f_agg)
     elif (config.robustness["on_single_weights"] == "yes") and (config.robustness["robustness_on"] == "yes"):
         # ONE RANDOMLY SAMPLED WEIGHT A TIME (MCDA runs (num_samples * num_indicators) times)
         iterative_random_weights_statistics: dict = _compute_scores_for_single_random_weight(
@@ -879,6 +880,7 @@ def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name
[…]
