+++ mcda/utils/utils_for_main.py
@@ -46,15 +46,14 @@ def check_config_error(condition: bool, error_message: str):
         raise ValueError(error_message)
 
 
-def check_config_setting(condition: bool, information_message: str,
-                         condition_robustness_on_weights: str, condition_robustness_on_indicators: str, mc_runs: int) \
-        -> (int, int):
+def check_config_setting(condition_robustness_on_weights: bool,
+                         condition_robustness_on_indicators: bool,
+                         mc_runs: int) -> (int, int):
     """
         Checks configuration settings and logs information messages.
 
         Parameters:
         - condition (bool): The condition to check.
-        - information_message (str): The information message to log.
 
         Returns:
         :rtype: Tuple[int, int]
@@ -63,7 +62,7 @@ def check_config_setting(condition: bool, information_message: str,
 
         Example:
         ```python
-        is_robustness_weights, is_robustness_indicators = check_config_setting(True, "Some information message",
+        is_robustness_weights, is_robustness_indicators = check_config_setting(True, "False,
                                                                                "True", "False", 1000)
         ```
         """
@@ -71,25 +70,28 @@ def check_config_setting(condition: bool, information_message: str,
     is_robustness_weights = 0
     is_robustness_indicators = 0
 
-    if condition:
-        logger.info(information_message)
+    if condition_robustness_on_weights:
+        logger.info("ProMCDA will consider uncertainty on the weights.")
         logger.info("Number of Monte Carlo runs: {}".format(mc_runs))
-    if str(condition) == condition_robustness_on_weights:
         is_robustness_weights = 1
-    elif str(condition) == condition_robustness_on_indicators:
+
+    elif condition_robustness_on_indicators:
+        logger.info("ProMCDA will consider uncertainty on the indicators.")
+        logger.info("Number of Monte Carlo runs: {}".format(mc_runs))
         is_robustness_indicators = 1
 
     return is_robustness_weights, is_robustness_indicators
 
 
-def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: pd.DataFrame,
+def process_indicators_and_weights(config: dict, input_matrix: pd.DataFrame,
                                    is_robustness_indicators: int, is_robustness_weights: int, polar: List[str],
-                                   mc_runs: int, num_indicators: int) -> Tuple[List[str], Union[list, List[list], dict]]:
+                                   mc_runs: int, num_indicators: int) -> Tuple[
+    List[str], Union[list, List[list], dict]]:
     """
     Process indicators and weights based on input parameters in the configuration.
 
     Parameters:
-    - input_matrix (DataFrame): The input matrix.
+    - input_matrix (DataFrame): The input matrix without alternatives.
     - is_robustness_indicators (int): Flag indicating whether the matrix should include indicator uncertainties
       (0 or 1).
     - is_robustness_weights (int): Flag indicating whether robustness analysis is considered for the weights
@@ -130,19 +132,19 @@ def process_indicators_and_weights(config: dict, input_matrix_no_alternatives: p
     :param polar: List[str]
     :param is_robustness_weights: int
     :param is_robustness_indicators: int
-    :param input_matrix_no_alternatives: pd.DataFrame
+    :param input_matrix: pd.DataFrame
     :param config: dict
     :param num_indicators: int
     """
 
-    num_unique = input_matrix_no_alternatives.nunique()
+    num_unique = input_matrix.nunique()
     cols_to_drop = num_unique[num_unique == 1].index
-    col_to_drop_indexes = input_matrix_no_alternatives.columns.get_indexer(cols_to_drop)
+    col_to_drop_indexes = input_matrix.columns.get_indexer(cols_to_drop)
 
[…]
