+++ mcda/mcda_run.py
@@ -102,30 +102,25 @@ def main(input_config: dict):
                            'Robustness analysis is requested: but on weights or indicators? Please clarify.')
 
         # Check seetings for robustness analysis on weights or indicators
-        condition_robustness_on_weights = ((config.robustness["on_single_weights"] == "yes" and
-                                            config.robustness["on_all_weights"] == "no" and
-                                            config.robustness["on_indicators"] == "no") or
-                                           (config.robustness["on_single_weights"] == "no" and
-                                            config.robustness["on_all_weights"] == "yes" and
-                                            config.robustness["on_indicators"] == "no"))
-        condition_robustness_on_indicators = (config.robustness["on_single_weights"] == "no" and
-                                              config.robustness["on_all_weights"] == "no" and
-                                              config.robustness["on_indicators"] == "yes")
+        condition_robustness_on_weights = (
+            (config.robustness['on_single_weights'] == 'yes' and
+            config.robustness['on_all_weights'] == 'no' and
+            config.robustness['on_indicators'] == 'no') or
+            (config.robustness['on_single_weights'] == 'no' and
+            config.robustness['on_all_weights'] == 'yes' and
+            config.robustness['on_indicators'] == 'no')
+        )
+        condition_robustness_on_indicators = (
+            (config.robustness['on_single_weights'] == 'no' and
+            config.robustness['on_all_weights'] == 'no' and
+            config.robustness['on_indicators'] == 'yes')
+        )
 
-        condition = condition_robustness_on_weights if condition_robustness_on_weights \
-                                                    else condition_robustness_on_indicators
 
         is_robustness_weights, is_robustness_indicators = \
-            check_config_setting(condition,
-                                 'ProMCDA will consider uncertainty on the weights',
-                                  str(condition_robustness_on_weights),str(condition_robustness_on_indicators),
+            check_config_setting(condition_robustness_on_weights,
+                                 condition_robustness_on_indicators,
                                   mc_runs)
-
-        is_robustness_weights, is_robustness_indicators = \
-            check_config_setting(condition,
-                                 'ProMCDA will consider uncertainty on the indicators',
-                                 str(condition_robustness_on_weights),str(condition_robustness_on_indicators),
-                                 mc_runs)
 
         marginal_pdf = config.monte_carlo_sampling["marginal_distribution_for_each_indicator"]
         logger.info("Read input matrix with uncertainty of the indicators at {}".format(
@@ -154,10 +149,10 @@ def main(input_config: dict):
     # If there is no uncertainty of the indicators:
     if is_robustness_indicators == 0:
         run_mcda_without_indicator_uncertainty(input_config, index_column_name, index_column_values,
-                                               input_matrix_no_alternatives, weights, f_norm, f_agg)
+                                               input_matrix_no_alternatives, weights, f_norm, f_agg, is_robustness_weights)
     # else (i.e. there is uncertainty):
     else:
-        run_mcda_with_indicator_uncertainty(input_config, input_matrix, index_column_name, index_column_values,
+        run_mcda_with_indicator_uncertainty(input_config, input_matrix_no_alternatives, index_column_name, index_column_values,
                                             mc_runs, is_sensitivity, f_agg, f_norm, weights, polar, marginal_pdf)
 
     logger.info(
