### 2024-02-29 5535a11
add contributing guidelines and reference in the readme
