### 2024-03-04 c8e76f5
fix toy example run
