+++ mcda/mcda_run.py
@@ -25,21 +25,21 @@ logger = logging.getLogger("ProMCDA")
 # noinspection PyTypeChecker
 def main(input_config: dict):
     """
-        Execute the ProMCDA (Probabilistic Multi-Criteria Decision Analysis) process.
+    Execute the ProMCDA (Probabilistic Multi-Criteria Decision Analysis) process.
 
-        Parameters:
-        - input_config (dictionary): Configuration parameters for the ProMCDA process.
+    Parameters:
+    - input_config (dictionary): Configuration parameters for the ProMCDA process.
 
-        Raises:
-        - ValueError: If there are issues with the input matrix, weights, or indicators.
+    Raises:
+    - ValueError: If there are issues with the input matrix, weights, or indicators.
 
-        This function performs the ProMCDA process based on the provided configuration.
-        It handles various aspects such as the sensitivity analysis and the robustness analysis.
-        The results are saved in output files, and plots are generated to visualize the scores and rankings.
+    This function performs the ProMCDA process based on the provided configuration.
+    It handles various aspects such as the sensitivity analysis and the robustness analysis.
+    The results are saved in output files, and plots are generated to visualize the scores and rankings.
 
-        Note: Ensure that the input matrix, weights, polarities and indicators (with or without uncertainty)
-        are correctly specified in the input configuration.
-        """
+    Note: Ensure that the input matrix, weights, polarities and indicators (with or without uncertainty)
+    are correctly specified in the input configuration.
+    """
 
     is_sensitivity = None
     is_robustness = None
