### 2024-03-07 c605747
add description of params and return in docstring
