### 2024-03-07 0b43beb
add selection of random seed for reproducibility
