+++ tests/unit_tests/test_utils_for_main.py
@@ -26,6 +26,7 @@ class TestUtils(unittest.TestCase):
             "monte_carlo_sampling": {
                 "monte_carlo_runs": 10000,
                 "num_cores": 1,
+                "random_seed": 42,
                 "marginal_distribution_for_each_indicator": ['exact', 'uniform', 'normal', 'poisson']},
             "output_path": "/path/to/output"
         }
