+++ tests/unit_tests/test_utils_for_parallelization.py
@@ -27,6 +27,7 @@ class TestUtilsForParallelization(unittest.TestCase):
             "monte_carlo_sampling": {
                 "monte_carlo_runs": 10000,
                 "num_cores": 1,
+                "random_seed": 42,
                 "marginal_distribution_for_each_indicator": ['exact', 'uniform', 'normal', 'exact', 'uniform']},
             "output_path": "/path/to/output"
         }
