+++ mcda/mcda_with_robustness.py
@@ -29,9 +29,11 @@ class MCDAWithRobustness:
 
     """
 
-    def __init__(self, config: Config, input_matrix: pd.DataFrame(), is_exact_pdf_mask=None, is_poisson_pdf_mask=None):
+    def __init__(self, config: Config, input_matrix: pd.DataFrame(), is_exact_pdf_mask=None, is_poisson_pdf_mask=None,
+                 random_seed=None):
         self.is_exact_pdf_mask = is_exact_pdf_mask
         self.is_poisson_pdf_mask = is_poisson_pdf_mask
+        self.random_seed = random_seed
         self._config = copy.deepcopy(config)
         self._input_matrix = copy.deepcopy(input_matrix)
 
@@ -87,8 +89,14 @@ class MCDAWithRobustness:
         input_matrix = self._input_matrix  # (AxnI)
         is_exact_pdf_mask = self.is_exact_pdf_mask
         is_poisson_pdf_mask = self.is_poisson_pdf_mask
+        random_seed = self.random_seed
 
-        np.random.seed(42)
+        default_random_seed = 42
+
+        if random_seed is not None:
+            np.random.seed(random_seed)
+        else:
+            np.random.seed(default_random_seed)
 
         sampled_matrices = []  # list long I
 

[…]
