### 2024-03-08 2dacf86
add missing docstrings to config, normalization, aggregation
