### 2024-03-11 49aafd0
add docstrings to utils for main and parallelization
