+++ mcda/utils/utils_for_parallelization.py
@@ -13,6 +13,38 @@ logger = logging.getLogger("ProMCDA utils for parallelization")
 
 
 def initialize_and_call_aggregation(args: Tuple[list, dict], method=None) -> pd.DataFrame:
+    """
+    Initialize an Aggregation object with given weights and call the aggregation method to calculate scores.
+
+    Parameters:
+    - args: a tuple containing a list of weights and a dictionary of data.
+    - method (optional): the aggregation method to use. Defaults to None.
+
+    Returns:
+    - pd.DataFrame: DataFrame containing scores calculated using the aggregation method.
+
+    Example:
+    ```python
+    weights = [0.2, 0.3, 0.5]
+    data = {'indicator1': [10, 20, 30], 'indicator2': [40, 50, 60], 'indicator3': [70, 80, 90]}
+
+    scores = initialize_and_call_aggregation((weights, data), method='weighted_sum')
+
+    scores
+    0   32.0
+    1   50.0
+    2   77.0
+
+    ```
+
+    This example initializes an Aggregation object with the given weights and calls the specified aggregation method
+    to calculate scores using the provided data. The resulting scores are returned as a DataFrame.
+
+    :param args: Tuple[list, dict]
+    :param method: str
+    :return scores_one_run: pd.DataFrame()
+    """
+
     weights, data = args
     agg = Aggregation(weights)
 
@@ -21,7 +53,42 @@ def initialize_and_call_aggregation(args: Tuple[list, dict], method=None) -> pd.
     return scores_one_run
 
 
-def initialize_and_call_normalization(args: Tuple[pd.DataFrame, list, str]) -> List[dict]:
+def initialize_and_call_normalization(args: Tuple[pd.DataFrame, list, str]) -> dict:
+    """
+    Initialize a Normalization object with given matrix and polarities, and call the normalization method to
+    calculate normalized indicators.
+
+    Parameters:
+    - args: a tuple containing a DataFrame of indicators, a list of polarities,
+      and a string specifying the normalization method.
+
+    Returns:
+    - dict: a dictionary containing normalized indicators.
+
+    Example:
+    ```python
+    matrix = pd.DataFrame({'indicator1': [10, 20, 30], 'indicator2': [40, 50, 60]})
+    polarities = ['+', '-', '+']
+    method = 'minmax'
+
+    normalized_indicators = initialize_and_call_normalization((matrix, polarities, method))
+
+    normalized_indicators:
+    {
+    'indicator1': [0.0, 0.5, 1.0],
+    'indicator2': [0.0, 0.5, 1.0]
+     }
+
+    ```
+
+    This example initializes a Normalization object with the given matrix and polarities, and calls the specified
+    normalization method to calculate normalized indicators. The resulting normalized indicators are returned
+    as a dictionary.
+
+    :param args: Tuple[pd.DataFrame(), List[str], str]
+    :return dict_normalized_matrix: dict
+    """
+
     matrix, polarities, method = args
     norm = Normalization(matrix, polarities)
 
@@ -31,6 +98,33 @@ def initialize_and_call_normalization(args: Tuple[pd.DataFrame, list, str]) -> L
 
 
 def normalize_indicators_in_parallel(norm: object, method=None) -> dict:
+    """
+        Normalize indicators in parallel using different normalization methods.
+
+        Parameters:
+        - norm: an object representing the normalization process.
+        - method: the normalization method to use. Options are 'minmax', 'target', 'standardized', or 'rank'.
+
+        Returns:
+        - A dictionary containing the normalized indicators for each method.
+
+        Example:
+        ```python
+        norm = Normalization(matrix, polarities)
+        normalized_indicators = normalize_indicators_in_parallel(norm, method='minmax')
+
+        ```
+
+        This example normalizes indicators in parallel using the min-max normalization method.
[…]
