+++ mcda/utils/utils_for_main.py
@@ -35,9 +35,9 @@ def check_config_error(condition: bool, error_message: str):
     Raises:
     - ValueError: If the condition is True, with the specified error message.
 
-    :return: None
     :param error_message: str
     :param condition: bool
+    :return: None
     """
 
     if condition:
@@ -47,24 +47,20 @@ def check_config_error(condition: bool, error_message: str):
 
 def check_config_setting(condition_robustness_on_weights: bool,
                          condition_robustness_on_indicators: bool,
-                         mc_runs: int) -> (int, int):
+                         mc_runs: int) \
+        -> (int, int):
     """
-        Checks configuration settings and logs information messages.
+    Checks configuration settings and logs information messages.
 
-        Parameters:
-        - condition (bool): The condition to check.
+    Return:
+    - is_robustness_weights, is_robustness_indicators, a tuple containing flags indicating robustness on weights and
+    indicators.
 
-        Returns:
-        :rtype: Tuple[int, int]
-        :return: is_robustness_weights, is_robustness_indicators, a tuple containing flags indicating robustness
-                 on weights and indicators.
+    :param condition_robustness_on_weights: bool
+    :param condition_robustness_on_indicators: bool
+    :rtype: Tuple[int, int]
 
-        Example:
-        ```python
-        is_robustness_weights, is_robustness_indicators = check_config_setting(True, "False,
-                                                                               "True", "False", 1000)
-        ```
-        """
+    """
 
     is_robustness_weights = 0
     is_robustness_indicators = 0
@@ -84,21 +80,20 @@ def check_config_setting(condition_robustness_on_weights: bool,
 
 def process_indicators_and_weights(config: dict, input_matrix: pd.DataFrame,
                                    is_robustness_indicators: int, is_robustness_weights: int, polar: List[str],
-                                   mc_runs: int, num_indicators: int) -> Tuple[
-    List[str], Union[list, List[list], dict]]:
+                                   mc_runs: int, num_indicators: int) \
+        -> Tuple[List[str], Union[list, List[list], dict]]:
     """
     Process indicators and weights based on input parameters in the configuration.
 
     Parameters:
-    - input_matrix (DataFrame): The input matrix without alternatives.
-    - is_robustness_indicators (int): Flag indicating whether the matrix should include indicator uncertainties
-      (0 or 1).
-    - is_robustness_weights (int): Flag indicating whether robustness analysis is considered for the weights
-      (0 or 1).
-    - marginal_pdf (list): List of marginal probability density functions for indicators.
-    - mc_runs (int): Number of Monte Carlo runs for robustness analysis.
-    - num_indicators: the number of indicators in the input matrix.
     - config: the configuration dictionary.
+    - input_matrix: the input matrix without alternatives.
+    - is_robustness_indicators: a flag indicating whether the matrix should include indicator uncertainties
+      (0 or 1).
+    - is_robustness_weights: a flag indicating whether robustness analysis is considered for the weights (0 or 1).
+    - marginal_pdf: a list of marginal probability density functions for indicators.
+    - mc_runs: number of Monte Carlo runs for robustness analysis.
+    - num_indicators: the number of indicators in the input matrix.
 
     Raises:
     - ValueError: If there are duplicated rows in the input matrix or if there is an issue with the configuration.
[…]
