+++ mcda/mcda_without_robustness.py
@@ -111,7 +111,7 @@ class MCDAWithoutRobustness:
         scores_harmonic = {}
         scores_minimum = {}
 
-        scores = pd.DataFrame
+        scores = pd.DataFrame()
         col_names_method = []
         col_names = ['ws-minmax_01', 'ws-target_01', 'ws-standardized_any', 'ws-rank',
                      'geom-minmax_without_zero', 'geom-target_without_zero', 'geom-standardized_without_zero', 'geom-rank',
