### 2024-03-15 wetransform-os/ProMCDA#27 @Flaminietta
Dear @B3J4y, in the main branch, latest commit e071d84, the toy example runs and finishes as expected for me.

If you can also try again and confirm that now everything is fine, I can then close this issue. Thank you!
