### 2024-03-20 5d31d40
remove relative paths and circular imports, introduce env variable for paths
