+++ mcda/utils/utils_for_main.py
@@ -1,19 +1,28 @@
+
+import os
 import argparse
 import json
 import pickle
 import random
-from typing import Union, Any, List
+import logging
+import sys
+from typing import Union, Any, List, Tuple
 from typing import Optional
 
+import numpy as np
 import pandas as pd
+from datetime import datetime
 from sklearn import preprocessing
 from sklearn.preprocessing import MinMaxScaler
 
+import mcda.utils.utils_for_parallelization as utils_for_parallelization
+import mcda.utils.utils_for_plotting as utils_for_plotting
 from mcda.configuration.config import Config
-from mcda.utils.utils_for_parallelization import *
-from mcda.utils.utils_for_plotting import *
 from mcda.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.mcda_with_robustness import MCDAWithRobustness
+
+DEFAULT_INPUT_DIRECTORY_PATH = './input_files'  # root directory of ProMCDA
+DEFAULT_OUTPUT_DIRECTORY_PATH = './output_files'
 
 log = logging.getLogger(__name__)
 
@@ -45,7 +54,6 @@ def check_config_error(condition: bool, error_message: str):
         raise ValueError(error_message)
 
 
-        
 def check_config_setting(condition_robustness_on_weights: bool, condition_robustness_on_indicators: bool, mc_runs: int,
                          random_seed: int) -> (int, int):
     """
@@ -160,9 +168,8 @@ def process_indicators_and_weights(config: dict, input_matrix: pd.DataFrame,
 def _handle_polarities_and_weights(is_robustness_indicators: int, is_robustness_weights: int, num_unique,
                                    col_to_drop_indexes: np.ndarray, polar: List[str], config: dict, mc_runs: int,
                                    num_indicators: int) \
-        -> Union[Tuple[List[str], list, None, None],
-        Tuple[List[str], None, List[List], None],
-        Tuple[List[str], None, None, dict]]:
+        -> Union[Tuple[List[str], list, None, None], Tuple[List[str], None, List[List], None],
+           Tuple[List[str], None, None, dict]]:
     """
     Manage polarities and weights based on the specified robustness settings, ensuring that the appropriate adjustments
     and normalizations are applied before returning the necessary data structures.
@@ -308,10 +315,34 @@ def check_input_matrix(input_matrix: pd.DataFrame) -> pd.DataFrame:
     return input_matrix_no_alternatives
 
 
+def ensure_directory_exists(path):
+    """
+    Ensure that the directory specified by the given path exists.
+    If the directory does not exist, create it and any intermediate directories as needed.
+
+    Parameters:
+        path (str): The path of the filr in the directory to ensure exists.
+
+    Example:
+    ```python
+    ensure_directory_exists(/path/to/directory/file.csv)
+    ```
+
+    :param path: str
+    :return: None
+    """
+    directory = os.path.dirname(path)
+    if not os.path.exists(directory):
+        os.makedirs(path)
+
+
 def read_matrix(input_matrix_path: str) -> pd.DataFrame:
     """
     Read an input matrix from a CSV file and return it as a DataFrame.
     Set the 'Alternatives' column as index column.
+
+    Note: a default path is assigned to the input_matrix_path in mcda_run.py, and it is used
+    unless a custom path is set in an environmental variable in the environment.
 
     Parameters:
     - input_matrix_path (str): path to the CSV file containing the input matrix.
@@ -323,9 +354,14 @@ def read_matrix(input_matrix_path: str) -> pd.DataFrame:
     :rtype: pd.DataFrame
     """
 
+    custom_input_path = os.environ.get('PROMCDA_INPUT_PATH')  # check if an environmental variable is set
[…]

[…]
