+++ mcda/utils/utils_for_main.py
@@ -21,8 +21,14 @@ from mcda.configuration.config import Config
 from mcda.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.mcda_with_robustness import MCDAWithRobustness
 
-DEFAULT_INPUT_DIRECTORY_PATH = './input_files'  # root directory of ProMCDA
-DEFAULT_OUTPUT_DIRECTORY_PATH = './output_files'
+DEFAULT_INPUT_DIRECTORY_PATH = './input_files'  # present in the root directory of ProMCDA
+DEFAULT_OUTPUT_DIRECTORY_PATH = './output_files' # present in the root directory of ProMCDA
+
+custom_input_path = os.environ.get('PROMCDA_INPUT_DIRECTORY_PATH')  # check if an environmental variable is set
+input_directory_path = custom_input_path if custom_input_path else DEFAULT_INPUT_DIRECTORY_PATH
+
+custom_output_path = os.environ.get('PROMCDA_OUTPUT_DIRECTORY_PATH')  # check if an environmental variable is set
+output_directory_path = custom_output_path if custom_output_path else DEFAULT_OUTPUT_DIRECTORY_PATH
 
 log = logging.getLogger(__name__)
 
@@ -145,7 +151,6 @@ def process_indicators_and_weights(config: dict, input_matrix: pd.DataFrame,
     :return: polar, norm_weights
     :rtype: Tuple[List[str], Union[List[list], dict]]
     """
-
     num_unique = input_matrix.nunique()
     cols_to_drop = num_unique[num_unique == 1].index
     col_to_drop_indexes = input_matrix.columns.get_indexer(cols_to_drop)
@@ -321,7 +326,7 @@ def ensure_directory_exists(path):
     If the directory does not exist, create it and any intermediate directories as needed.
 
     Parameters:
-        path (str): The path of the filr in the directory to ensure exists.
+        path (str): The path of the file in the directory to ensure exists.
 
     Example:
     ```python
@@ -331,9 +336,12 @@ def ensure_directory_exists(path):
     :param path: str
     :return: None
     """
-    directory = os.path.dirname(path)
-    if not os.path.exists(directory):
-        os.makedirs(path)
+    try:
+        directory = os.path.dirname(path)
+        if not os.path.exists(directory):
+            os.makedirs(directory)
+    except Exception as e:
+        logging.error(f"An error occurred while ensuring directory exists: {e}")
 
 
 def read_matrix(input_matrix_path: str) -> pd.DataFrame:
@@ -353,15 +361,15 @@ def read_matrix(input_matrix_path: str) -> pd.DataFrame:
     :param input_matrix_path: str
     :rtype: pd.DataFrame
     """
-
-    custom_input_path = os.environ.get('PROMCDA_INPUT_PATH')  # check if an environmental variable is set
-    input_directory_path = custom_input_path if custom_input_path else DEFAULT_INPUT_DIRECTORY_PATH
-
-    ensure_directory_exists(input_directory_path)
+    try:
+        ensure_directory_exists(input_directory_path)
+    except IOError:
+        logging.error(f"Reading a matrix failed.")
 
     try:
         full_file_path = os.path.join(input_directory_path, input_matrix_path)
         with open(full_file_path, 'r') as fp:
+            logger.info("Reading the input matrix in {}".format(full_file_path))
             matrix = pd.read_csv(fp, sep="[,;:]", decimal='.', engine='python')
             data_types = {col: 'float64' for col in matrix.columns[1:]}
             matrix = matrix.astype(data_types)
@@ -414,7 +422,6 @@ def parse_args():
 
     :rtype: str
     """
-
     parser = argparse.ArgumentParser()
[…]
