### 2024-05-08 81d26fd
Update README based on the comments on Issue #30
