### 2024-05-08 b075326
Update README based on the comments on Issue #30
