### 2024-05-15 59e4f76
Update JOSS paper based on minor comments provided in Issue #29
