+++ JOSS_paper/paper.md
@@ -1,98 +1,0 @@
----
-title: "ProMCDA: A Python package for Probabilistic Multi-Criteria Decision Analysis"
-tags:
-- Python
-- Multi-criteria Decision Analysis (MCDA)
-- decision support
-- probabilistic approach
-- sensitivity analysis
-- Monte Carlo sampling
-date: "15 November 2023"
-output:
-  html_document: 
-    fig_caption: yes
-  pdf_document: 
-    fig_caption: yes
-authors:
-- name: Flaminia Catalli
-  orcid: "0000-0003-0515-5282"
-  equal-contrib: yes
-  affiliation: 1
-- name: Matteo Spada
-  orcid: "0000-0001-9265-9491"
-  equal-contrib: yes
-  affiliation: 2
-bibliography: paper.bib
-affiliations:
-- name: wetransform GmbH, Germany
-  index: 1
-- name: Zürich University of Applied Sciences, School of Engineering, INE Institute
-    of Sustainable Development, Switzerland
-  index: 2
-editor_options:
-  markdown:
-  chunk_output_type: inline
----
-
-# Summary
-
-Multi-Criteria Decision Analysis (MCDA) is a formal process to assist decision makers (DMs) in structuring their decision problems and to provide them with tools and methods leading to recommendations on the decisions at stake (@roy_decision_1996). The recommendations are based on a comprehensive identification of the alternatives considered and the selection of criteria/subcriteria/etc. to evaluate them, which are aggregated taking into account the preferences of the DMs (@bouyssou_problem_2006). In the literature, there is a wide range of MCDA methods used to integrate information and either classify alternatives into preference classes or rank them from best to worst (@cinelli_proper_2022). In the context of ranking and benchmarking alternatives across complex concepts, composite indicators (CIs) are the most widely used synthetic measures (@greco_methodological_2019). Indeed, they have been applied, for example, in the context of environmental quality (@otoiu_proposing_2018), resilience of energy supply (@gasser_comprehensive_2020), sustainability (@volkart_interdisciplinary_2016), global competitiveness (@klaus_schwab_global_2018), etc. However, the uncertainty of the criteria, the effect of tuning the weights relative to them, and the choice of methods (normalization/aggregation) to construct CIs, have been shown to influence the final ranking of alternatives (e.g. @cinelli_mcda_2020).
-
-The `ProMCDA` Python module proposed here allows a DM to explore the sensitivity and robustness of the CIs results in a user-friendly way. In other words, it allows the user to assess either the sensitivity related to the choice of normalization and/or aggregation method, but also to account for uncertainty in the criteria and weights.
-
-# Statement of need
-
-There are already dedicated tools for estimating CIs in the literature. In *R*, there is an existing package called `COINr`, which allows the user to develop CIs by including all common operations, from criteria selection, data treatment, normalization and aggregation, and sensitivity analysis (@becker_coinr_2022). There are other packages in R, such as `compind`, that focus on weighting and aggregation (@fusco_spatial_2018). In *MATLAB*, there are some packages dedicated to specific parts of CI development, such as the `CIAO` tool (@linden_framework_2021). The Python module `Decisi-o-Rama` (@chacon-hurtado_decisi-o-rama_2021) focuses on the implementation of the Multi-Attribute Utility Theory (MAUT) to normalize criteria, considering a hierarchical criteria structure and uncertain criteria, and to aggregate the results using different aggregation methods. Finally, the web tool called `MCDA Index Tool` allows sensitivity analysis based on different combinations of normalization functions and aggregation methods ([MCDA Index Tool](https://www.mcdaindextool.net)).
-
[…]
