### 2024-05-15 9d02d09
Update JOSS paper based on minor comments provided in Issue #29
