### 2024-05-15 22eac23
Update JOSS paper based on minor comments provided in Issue #29 - Missing files
