### 2024-05-15 e35072b
Update JOSS paper based on minor comments provided in Issue #29 - Missing files
