### 2024-05-15 wetransform-os/ProMCDA#29 @mspada
@limengbinggz I just updated the paper including your minor comments. I'll now merge the branch to the main.
