### 2024-06-13 #6190 @mspada
Hi @galessiorob,

@Flaminietta and I closed some of the open issues so far, the question is how does all the process proceed? I saw that, for example, @paulrougieux generated the PDF for review on the 22.04 before the updated version of the paper, which included some of the issues raised by @limengbinggz and @B3J4y.
