### 2024-06-27 wetransform-os/ProMCDA#27 @B3J4y
Sorry, the problem is still there:
![ProMCMA](https://github.com/wetransform-os/ProMCDA/assets/1398738/0870a3fb-885f-4ed2-a715-5af36e351527)

I left it open for a while to look if there was a calculation which needs some time, but this process doesn't come back.

I use windows.
