### 2024-07-02 wetransform-os/ProMCDA#27 @mspada
@B3J4y I run the test on windows at it works smooth. Just a question: did the figures open on your browser or not? If not it might be a problem with some close ports that do not allow plotly to open the figures and, therefore, not ending the process. Let us know!
