### 2024-07-08 wetransform-os/ProMCDA#27 @B3J4y
There are 2 figures. I append the open images:
![Figure1](https://github.com/wetransform-os/ProMCDA/assets/1398738/e77353c4-7c9d-4bb5-880b-b682003714e7)
![Figure2](https://github.com/wetransform-os/ProMCDA/assets/1398738/53ae44d7-95ff-40b6-b93e-e763a2be5006)
