### 2024-07-08 wetransform-os/ProMCDA#27 @mspada
@B3J4y Thank you! Therefore, there might be some issue with the saving of some of the output files.  When you run it, is it saving some output files in the output directory or not? If yes, what output files is ProMCDA saving?
