### 2024-07-09 wetransform-os/ProMCDA#27 @B3J4y
You're welcome. It is tough to fix bugs remotely on another machine without having direct access. Hopefully, we can manage this.

Here is an image of the output folder:

![Output_files](https://github.com/wetransform-os/ProMCDA/assets/1398738/45c331c2-11fc-4985-bfa7-8b94e6eec301)
