### 2024-07-09 wetransform-os/ProMCDA#27 @mspada
Perfect, thanks! I know this is quite a tough way to try to fix an issue. In any case, it looks like that the issue is related to the figures, probably in the saving of them in the output folder. I'll have a look at it. I'll keep you posted
