### 2024-07-10 wetransform-os/ProMCDA#27 @mspada
@B3J4y I guess I found the issue. It looks like that it is related to kaleido, which is used to save the static image of the plotly plots. In some cases (only on windows and not for everyone) the new version of kaleido (0.2.1), which is used in ProMCDA, is not correctly working. To solve this the community propose to downgrade kaleido to the version 0.1.0.post1 as discussed in https://community.plotly.com/t/static-image-export-hangs-using-kaleido/61519 and https://github.com/plotly/Kaleido/issues/126
Could you please try to downgrade kaleido to the version 0.1.0.post1 and run the test once again?
