### 2024-07-23 wetransform-os/ProMCDA#27 @B3J4y
@mspada Thank you for finding the solution. I haven't had the time to downgrade Kaleido, and I feel that whether it works or not, your package will be very weak in terms of reusability if it relies on this mechanism for your Windows users. Could you possibly package it into a Docker container and add a Dockerfile to the repository? This could solve the Windows problems and provide users with a system-independent way to run the software.
