+++ requirements.txt
@@ -8,6 +8,5 @@ pytest
 scikit-learn
 scipy
 Pillow
-kaleido
 
 
