### 2024-07-29 c75bedc
correct typos in README
