+++ requirements.txt
@@ -8,5 +8,6 @@ pytest
 scikit-learn
 scipy
 Pillow
+kaleido
 
 
