### 2024-07-29 b8cd85b
build: add dockerfile and .dockerignore
