### 2024-07-29 26f0185
build: add dockerfile and .dockerignore
