### 2024-07-30 wetransform-os/ProMCDA#27 @Flaminietta
Dear @B3J4y, your request is addressed in [this PR](https://github.com/wetransform-os/ProMCDA/pull/46) (ProMCDA is now also dockerized).
