### 2024-07-31 #6190 @Flaminietta
Dear all, dear @galessiorob,

with the [last PR](https://github.com/wetransform-os/ProMCDA/pull/46), which solves the potential problem for some Windows users by dockerizing ProMCDA, I think that we have now addressed all issues related to this review.

Best regards,
Flaminia
