### 2024-08-01 #6190 @B3J4y
On my end, everything has been checked, the software works out of the box, and this lengthy process is now complete.

Thank you, @Flaminietta, for your effort and for pushing the development forward.

Thank you, @galessiorob, for consistently organizing and reminding everyone about their tasks throughout this long period.
