### 2024-08-30 #6190 @mspada
@galessiorob I'm writing you since it is quite some time that we didn't hear anything anymore about the status of the paper. I guess the problem is that @paulrougieux didn't finalise his review yet, or?
Looking forward for your update.
Best,
@mspada and @Flaminietta
