+++ JOSS_paper/paper.md
@@ -36,27 +36,28 @@ editor_options:
 
 # Summary
 
-Multi-Criteria Decision Analysis (MCDA) is a formal process to assist decision makers in structuring their decision problems and to provide them with tools and methods leading to recommendations on the decisions at stake (@roy_decision_1996). The recommendations are based on a comprehensive identification of the alternatives considered and the selection of criteria/subcriteria/etc. to evaluate them, which are aggregated taking into account the preferences of the decision makers (@bouyssou_problem_2006). In the literature, there is a wide range of MCDA methods used to integrate information and either classify alternatives into preference classes or rank them from best to worst (@cinelli_proper_2022). In the context of ranking and benchmarking alternatives across complex concepts, composite indicators (CIs) are the most widely used synthetic measures (@greco_methodological_2019). Indeed, they have been applied, for example, in the context of environmental quality (@otoiu_proposing_2018), resilience of energy supply (@gasser_comprehensive_2020), sustainability (@volkart_interdisciplinary_2016), global competitiveness (@klaus_schwab_global_2018), etc. However, the uncertainty of the criteria, the effect of tuning the weights relative to them, and the choice of methods (normalization/aggregation) to construct CIs, have been shown to influence the final ranking of alternatives (e.g. @cinelli_mcda_2020).
+Multi-Criteria Decision Analysis (MCDA) is a formal process used to assist decision-makers in structuring complex decision problems and providing recommendations based on a comprehensive evaluation of alternatives. This evaluation is conducted by selecting relevant criteria and subcriteria, which are then aggregated according to the preferences of the decision-makers to produce a ranking or classification of the alternatives (@roy_decision_1996; @bouyssou_problem_2006). A wide range of MCDA methods are available in the literature for integrating information to classify alternatives into preference classes or rank them from best to worst (@cinelli_proper_2022). Among these, composite indicators (CIs) are commonly used synthetic measures for ranking and benchmarking alternatives across complex concepts (@greco_methodological_2019). Examples of CI applications include environmental quality assessment (@otoiu_proposing_2018), resilience of energy supply (@gasser_comprehensive_2020), sustainability (@volkart_interdisciplinary_2016), and global competitiveness (@klaus_schwab_global_2018).
 
-The `ProMCDA` Python module proposed here allows a decision maker to explore the sensitivity and robustness of the CIs results in a user-friendly way. In other words, it allows the user to assess either the sensitivity related to the choice of normalization and/or aggregation method, but also to account for uncertainty in the criteria and weights.
+However, the final ranking of alternatives in MCDA can be influenced by various factors such as uncertainty in the criteria, the choice of weights assigned to them, and the selection of methods for normalization and aggregation to construct CIs (@cinelli_mcda_2020). To address these challenges, the `ProMCDA` Python module has been developed to allow decision-makers to explore the sensitivity and robustness of CI results in a user-friendly manner. This tool facilitates sensitivity analysis related to the choice of normalization and aggregation methods and accounts for uncertainty in criteria and weights, providing a systematic approach to understanding the impact of these factors on decision outcomes.
 
 # Statement of need
 
[…]
