### 2024-09-03 #6190 @galessiorob
👋 @mspada @Flaminietta  apologies for the delay! Summer got in the way, but I think this is pretty close to ready for publishing! A few things before we do that:
* I checked out a [PR](https://github.com/wetransform-os/ProMCDA/pull/47) with some edit suggestions for the paper, feel free to adopt as much as you like
* @paulrougieux given how long this paper has been in the editing queue and the fact that we have two complete reviews (thank you @B3J4y and @limengbinggz) I have removed you from the reviewers list, but please feel free to volunteer to review upcoming papers that would benefit from your expertise
* Once you review the paper suggestions and we have the final version, we'll need you to deposit your software in Zenodo and post the DOI here and the most recent version. After that, we can set the final version, generate a final proof, and have an Editor in Chief do the publishing.

So close! I'll check in soon, thanks again for lending your expertise @B3J4y @limengbinggz
