### 2024-09-04 5790d7d
refactor: use model classes and also refactor original config implementation to use model classes
