+++ models/sensitivity_pro_mcda.py
@@ -1,0 +1,145 @@
+# coding: utf-8
+
+from __future__ import absolute_import
+from datetime import date, datetime  # noqa: F401
+
+from typing import List, Dict  # noqa: F401
+
+from swagger_server.models.base_model_ import Model
+from swagger_server.models.sensitivity_ranked_alternatives import SensitivityRankedAlternatives  # noqa: F401,E501
+from swagger_server import util
+
+
+class SensitivityProMCDA(Model):
+    """NOTE: This class is auto generated by the swagger code generator program.
+
+    Do not edit the class manually.
+    """
+    def __init__(self, name: str=None, raw_scores: SensitivityRankedAlternatives=None, avg_scores_std: SensitivityRankedAlternatives=None, normalized_scores: SensitivityRankedAlternatives=None):  # noqa: E501
+        """SensitivityProMCDA - a model defined in Swagger
+
+        :param name: The name of this SensitivityProMCDA.  # noqa: E501
+        :type name: str
+        :param raw_scores: The raw_scores of this SensitivityProMCDA.  # noqa: E501
+        :type raw_scores: SensitivityRankedAlternatives
+        :param avg_scores_std: The avg_scores_std of this SensitivityProMCDA.  # noqa: E501
+        :type avg_scores_std: SensitivityRankedAlternatives
+        :param normalized_scores: The normalized_scores of this SensitivityProMCDA.  # noqa: E501
+        :type normalized_scores: SensitivityRankedAlternatives
+        """
+        self.swagger_types = {
+            'name': str,
+            'raw_scores': SensitivityRankedAlternatives,
+            'avg_scores_std': SensitivityRankedAlternatives,
+            'normalized_scores': SensitivityRankedAlternatives
+        }
+
+        self.attribute_map = {
+            'name': 'name',
+            'raw_scores': 'rawScores',
+            'avg_scores_std': 'avgScoresStd',
+            'normalized_scores': 'normalizedScores'
+        }
+        self._name = name
+        self._raw_scores = raw_scores
+        self._avg_scores_std = avg_scores_std
+        self._normalized_scores = normalized_scores
+
+    @classmethod
+    def from_dict(cls, dikt) -> 'SensitivityProMCDA':
+        """Returns the dict as a model
+
+        :param dikt: A dict.
+        :type: dict
+        :return: The SensitivityProMCDA of this SensitivityProMCDA.  # noqa: E501
+        :rtype: SensitivityProMCDA
+        """
+        return util.deserialize_model(dikt, cls)
+
+    @property
+    def name(self) -> str:
+        """Gets the name of this SensitivityProMCDA.
+
+        The name of the alternatives.  # noqa: E501
+
+        :return: The name of this SensitivityProMCDA.
+        :rtype: str
+        """
+        return self._name
+
+    @name.setter
+    def name(self, name: str):
+        """Sets the name of this SensitivityProMCDA.
+
+        The name of the alternatives.  # noqa: E501
+
+        :param name: The name of this SensitivityProMCDA.
+        :type name: str
+        """
+
+        self._name = name
+
+    @property
+    def raw_scores(self) -> SensitivityRankedAlternatives:
+        """Gets the raw_scores of this SensitivityProMCDA.
+
+
+        :return: The raw_scores of this SensitivityProMCDA.
+        :rtype: SensitivityRankedAlternatives
+        """
+        return self._raw_scores
+
+    @raw_scores.setter
+    def raw_scores(self, raw_scores: SensitivityRankedAlternatives):
+        """Sets the raw_scores of this SensitivityProMCDA.
+
+
+        :param raw_scores: The raw_scores of this SensitivityProMCDA.
[…]
