### 2024-09-05 6694724
docs: update paper.bib
