+++ JOSS_paper/paper.bib
@@ -68,6 +68,7 @@
   volume={22},
   pages={101368},
   year={2023},
+  doi = {10.1016/j.softx.2023.101368},
   publisher={Elsevier}
 }
 
