### 2024-09-05 #6190 @mspada
Hi @galessiorob,

- we just checked the last version of the manuscript and  made a correction in the bibliography by adding a missing doi. Now the paper is fine from our side and all the different links (incl. ORCID) and affiliations are correct.
- We made the release of ProMCDA as requested. The release version number is v1.1.0
- we archive the release to Zenodo as requested. All the provided info are in line with the ones on the paper. The Zenodo doi is https://doi.org/10.5281/zenodo.13694286

Let us know if you need something else from us.
Thank you for the support
@mspada and @Flaminietta
