### 2024-09-05 #6190 @galessiorob
@mspada and @Flaminietta thanks for the swift changes! Could you check that last DOI that is showing up as invalid, please? Just remove the "https://doi.org/" prefix

And the "The global competitiveness report 2018", could you add it too, please?
