### 2024-09-05 #6190 @mspada
@galessiorob for the "The global competitiveness report 2018" there is no doi, but just an ISBN, should I added it or not?
