### 2024-09-05 #6190 @galessiorob
> Do we need to create an additional Release after updating the paper with the doi problem?

I don't believe so, but will ask the Editor in Chief

> "The global competitiveness report 2018" there is no doi, but just an ISBN, should I added it or not?

Yes, let's add it please, and we can also ask the Editor, we'll leave a note. Thanks!
