### 2024-09-06 b8a26ca
Update paper.bib
Correct a citation.
