+++ JOSS_paper/paper.bib
@@ -11,7 +11,7 @@
 
 @article{pereira_enhancing_2024,
 	title = {Enhancing Decision Analysis with a Large Language Model: {pyDecision} a Comprehensive Library of {MCDA} Methods in Python},
-	doi = {https://doi.org/10.48550/arXiv.2404.06370},
+	doi = {10.48550/arXiv.2404.06370},
 	abstract = {Purpose: Multicriteria decision analysis ({MCDA}) has become increasingly essential for decision-making in complex environments. In response to this need, the {pyDecision} library, implemented in Python and available at https://bit.ly/3tLFGtH, has been developed to provide a comprehensive and accessible collection of {MCDA} methods. Methods: The {pyDecision} offers 70 {MCDA} methods, including {AHP}, {TOPSIS}, and the {PROMETHEE} and {ELECTRE} families. Beyond offering a vast range of techniques, the library provides visualization tools for more intuitive results interpretation. In addition to these features, {pyDecision} has integrated {ChatGPT}, an advanced Large Language Model, where decision-makers can use {ChatGPT} to discuss and compare the outcomes of different methods, providing a more interactive and intuitive understanding of the solutions. Findings: Large Language Models are undeniably potent but can sometimes be a double-edged sword. Its answers may be misleading without rigorous verification of its outputs, especially for researchers lacking deep domain expertise. It's imperative to approach its insights with a discerning eye and a solid foundation in the relevant field. Originality: With the integration of {MCDA} methods and {ChatGPT}, {pyDecision} is a significant contribution to the scientific community, as it is an invaluable resource for researchers, practitioners, and decision-makers navigating complex decisionmaking problems and seeking the most appropriate solutions based on {MCDA} methods.},
 	journaltitle = {Arvix},
 	author = {Pereira, Valdecy and Basilio, Marcio Pereira and Santos, Carlos Henrique Tarjano},
