+++ JOSS_paper/paper.bib
@@ -172,6 +172,7 @@
 @article{klaus_schwab_global_2018,
 	title = {The global competitiveness report 2018},
 	url = {https://apo.org.au/node/262576},
+        isbn = {978-92-95044-76-0},
 	abstract = {The 2018 edition of the Global Competitiveness Report represents a milestone in the four-decade history of the series, with the introduction of the new Global Competitiveness Index 4.0. The new index sheds light on an emerging set of drivers of productivity and long-term growth in the era of the Fourth Industrial Revolution.},
 	author = {{Klaus Schwab}},
 	year = {2018},
