### 2024-09-06 #6190 @mspada
Hi @galessiorob,
we removed the "https://doi.org/" prefix where requested and added the ISBN to the report.
Should we generate a new tag?
Thanks
