### 2024-09-06 ef9614b
Remove spaces
In XML, IDs can't have spaces, the space in the anchor is likely causing the reference issue in the JATS file. This removes the space in both the markdown anchor and the corresponding ID in the JATS file.
