### 2024-09-06 #6190 @galessiorob
Found the issue with the table ids pointed above, here's a [PR](https://github.com/wetransform-os/ProMCDA/pull/50) that removes the spaces. I think this should fix it. @mspada  @Flaminietta
