### 2024-09-10 a809166
refactor: refer ProMCDA models instead of ranking-service models
