+++ mcda/models/inline_response2001.py
@@ -5,8 +5,8 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server import util
+from ProMCDA.mcda.models.base_model_ import Model
+from ProMCDA.mcda import util
 
 
 class InlineResponse2001(Model):
