+++ mcda/models/configuration_sensitivity.py
@@ -5,23 +5,23 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server import util
+from ProMCDA.mcda.models.base_model_ import Model
+from ProMCDA.mcda import util
 
 
-class ConfigurationModifiedInputSensitivity(Model):
+class ConfigurationSensitivity(Model):
     """NOTE: This class is auto generated by the swagger code generator program.
 
     Do not edit the class manually.
     """
     def __init__(self, sensitivity_on: str=None, normalization: str=None, aggregation: str=None):  # noqa: E501
-        """ConfigurationModifiedInputSensitivity - a model defined in Swagger
+        """ConfigurationSensitivity - a model defined in Swagger
 
-        :param sensitivity_on: The sensitivity_on of this ConfigurationModifiedInputSensitivity.  # noqa: E501
+        :param sensitivity_on: The sensitivity_on of this ConfigurationSensitivity.  # noqa: E501
         :type sensitivity_on: str
-        :param normalization: The normalization of this ConfigurationModifiedInputSensitivity.  # noqa: E501
+        :param normalization: The normalization of this ConfigurationSensitivity.  # noqa: E501
         :type normalization: str
-        :param aggregation: The aggregation of this ConfigurationModifiedInputSensitivity.  # noqa: E501
+        :param aggregation: The aggregation of this ConfigurationSensitivity.  # noqa: E501
         :type aggregation: str
         """
         self.swagger_types = {
@@ -40,34 +40,34 @@ class ConfigurationModifiedInputSensitivity(Model):
         self._aggregation = aggregation
 
     @classmethod
-    def from_dict(cls, dikt) -> 'ConfigurationModifiedInputSensitivity':
+    def from_dict(cls, dikt) -> 'ConfigurationSensitivity':
         """Returns the dict as a model
 
         :param dikt: A dict.
         :type: dict
-        :return: The ConfigurationModifiedInput_sensitivity of this ConfigurationModifiedInputSensitivity.  # noqa: E501
-        :rtype: ConfigurationModifiedInputSensitivity
+        :return: The Configuration_sensitivity of this ConfigurationSensitivity.  # noqa: E501
+        :rtype: ConfigurationSensitivity
         """
         return util.deserialize_model(dikt, cls)
 
     @property
     def sensitivity_on(self) -> str:
-        """Gets the sensitivity_on of this ConfigurationModifiedInputSensitivity.
+        """Gets the sensitivity_on of this ConfigurationSensitivity.
 
         A sensitivity analysis is performed ['yes'] or not ['no'].  # noqa: E501
 
-        :return: The sensitivity_on of this ConfigurationModifiedInputSensitivity.
+        :return: The sensitivity_on of this ConfigurationSensitivity.
         :rtype: str
         """
         return self._sensitivity_on
 
     @sensitivity_on.setter
     def sensitivity_on(self, sensitivity_on: str):
-        """Sets the sensitivity_on of this ConfigurationModifiedInputSensitivity.
+        """Sets the sensitivity_on of this ConfigurationSensitivity.
 
         A sensitivity analysis is performed ['yes'] or not ['no'].  # noqa: E501
 
-        :param sensitivity_on: The sensitivity_on of this ConfigurationModifiedInputSensitivity.
+        :param sensitivity_on: The sensitivity_on of this ConfigurationSensitivity.
         :type sensitivity_on: str
         """
         allowed_values = ["yes", "no"]  # noqa: E501
@@ -81,22 +81,22 @@ class ConfigurationModifiedInputSensitivity(Model):
 
     @property
[…]
