+++ mcda/models/__init__.py
@@ -1,0 +1,6 @@
+# coding: utf-8
+
+# flake8: noqa
+from __future__ import absolute_import
+# import models into model package
+from ProMCDA.mcda.models.log_normal_indicator_values import LogNormalIndicatorValues
