+++ models/__init__.py
@@ -1,25 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-from __future__ import absolute_import
-# import models into model package
-from swagger_server.models.configuration import Configuration
-from swagger_server.models.configuration_monte_carlo_sampling import ConfigurationMonteCarloSampling
-from swagger_server.models.configuration_robustness import ConfigurationRobustness
-from swagger_server.models.configuration_sensitivity import ConfigurationSensitivity
-from swagger_server.models.convert_csv_to_json_body import ConvertCsvToJsonBody
-from swagger_server.models.error import Error
-from swagger_server.models.exact_indicator_values import ExactIndicatorValues
-from swagger_server.models.inline_response200 import InlineResponse200
-from swagger_server.models.inline_response2001 import InlineResponse2001
-from swagger_server.models.inline_response2002 import InlineResponse2002
-from swagger_server.models.log_normal_indicator_values import LogNormalIndicatorValues
-from swagger_server.models.normal_indicator_values import NormalIndicatorValues
-from swagger_server.models.poisson_indicator_values import PoissonIndicatorValues
-from swagger_server.models.sensitivity_mcda import SensitivityMCDA
-from swagger_server.models.sensitivity_pro_mcda import SensitivityProMCDA
-from swagger_server.models.sensitivity_ranked_alternatives import SensitivityRankedAlternatives
-from swagger_server.models.sensitivity_ranked_alternatives_inner import SensitivityRankedAlternativesInner
-from swagger_server.models.simple_mcda import SimpleMCDA
-from swagger_server.models.simple_pro_mcda import SimpleProMCDA
-from swagger_server.models.uniform_indicator_values import UniformIndicatorValues
