+++ mcda/models/sensitivity_pro_mcda.py
@@ -5,9 +5,9 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server.models.sensitivity_ranked_alternatives import SensitivityRankedAlternatives  # noqa: F401,E501
-from swagger_server import util
+from ProMCDA.mcda.models.base_model_ import Model
+from ProMCDA.mcda.models.sensitivity_ranked_alternatives import SensitivityRankedAlternatives  # noqa: F401,E501
+from ProMCDA.mcda import util
 
 
 class SensitivityProMCDA(Model):

[…]
