### 2024-09-10 02a26a5
fix: tests to import correct Configuration object
