### 2024-09-18 #6190 @galessiorob
@oliviaguest gentle ping on this 🙏  I haven't found references on how to fix the table tags in the XML file (or maybe @arfon you've run into a case like this?)
