### 2024-09-24 #6190 @mspada
@galessiorob I found a similar error during the preparation of the paper acceptance here: https://github.com/openjournals/joss-reviews/issues/4622
it looks like they solved this way: https://github.com/bancaditalia/black-it/pull/21/files
