### 2024-09-24 a5f92f9
refactor: update to run ProMCDA from cli
