+++ requirements.txt
@@ -9,5 +9,10 @@ scikit-learn
 scipy
 Pillow
 kaleido
+connexion >= 2.6.0
+connexion[swagger-ui] >= 2.6.0
+python_dateutil == 2.6.0
+setuptools >= 21.0.0
+swagger-ui-bundle >= 0.0.2
 
 
