### 2024-09-30 26c637b
refactor: address review comments and add doc strings
