+++ mcda/models/__init__.py
@@ -2,5 +2,3 @@
 
 # flake8: noqa
 from __future__ import absolute_import
-# import models into model package
-from ProMCDA.mcda.models.log_normal_indicator_values import LogNormalIndicatorValues
