+++ mcda/models/base_model_.py
@@ -1,3 +1,9 @@
+"""
+The class Model is designed to represent a data model with serialization and deserialization capabilities.
+It provides a framework for representing data models that can be easily serialized to and from dictionaries.
+It supports operations for equality checks, and for creating more complex models in Python applications with data exchange.
+"""
+
 import pprint
 
 import six

[…]
