### 2024-10-08 eae5fec
refactor:start adding new logic for library
