+++ mcda/utils/utils_for_main.py
@@ -13,7 +13,6 @@ import numpy as np
 import pandas as pd
 from datetime import datetime
 from sklearn import preprocessing
-from sklearn.preprocessing import MinMaxScaler
 
 import mcda.utils.utils_for_parallelization as utils_for_parallelization
 import mcda.utils.utils_for_plotting as utils_for_plotting
@@ -37,287 +36,6 @@ logging.getLogger('PIL').setLevel(logging.WARNING)
 FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
 logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
 logger = logging.getLogger("ProMCDA")
-
-
-def check_config_error(condition: bool, error_message: str):
-    """
-    Check a condition and raise a ValueError with a specified error message if the condition is True.
-
-    Parameters:
-    - condition (bool): The condition to check.
-    - error_message (str): The error message to raise if the condition is True.
-
-    Raises:
-    - ValueError: If the condition is True, with the specified error message.
-
-    :param error_message: str
-    :param condition: bool
-    :return: None
-    """
-
-    if condition:
-        logger.error('Error Message', stack_info=True)
-        raise ValueError(error_message)
-
-
-def check_config_setting(condition_robustness_on_weights: bool, condition_robustness_on_indicators: bool, mc_runs: int,
-                         random_seed: int) -> (int, int):
-    """
-    Checks configuration settings and logs information messages.
-
-    Returns:
-    - is_robustness_weights, is_robustness_indicators, booleans indicating if robustness is considered
-    on weights or indicators.
-
-    Example:
-    ```python
-    is_robustness_weights, is_robustness_indicators = check_config_setting(True, False, 1000, 42)
-    ```
-
-    :param condition_robustness_on_weights: bool
-    :param condition_robustness_on_indicators: bool
-    :param mc_runs: int
-    :param random_seed: int
-    :return: (is_robustness_weights, is_robustness_indicators)
-    :rtype: Tuple[int, int]
-    """
-    is_robustness_weights = 0
-    is_robustness_indicators = 0
-
-    if condition_robustness_on_weights:
-        logger.info("ProMCDA will consider uncertainty on the weights.")
-        logger.info("Number of Monte Carlo runs: {}".format(mc_runs))
-        logger.info("The random seed used is: {}".format(random_seed))
-        is_robustness_weights = 1
-
-    elif condition_robustness_on_indicators:
-        logger.info("ProMCDA will consider uncertainty on the indicators.")
-        logger.info("Number of Monte Carlo runs: {}".format(mc_runs))
-        logger.info("The random seed used is: {}".format(random_seed))
-        is_robustness_indicators = 1
-
-    return is_robustness_weights, is_robustness_indicators
-
-
-def process_indicators_and_weights(config: dict, input_matrix: pd.DataFrame,
-                                   is_robustness_indicators: int, is_robustness_weights: int, polar: List[str],
-                                   mc_runs: int, num_indicators: int) \
-        -> Tuple[List[str], Union[list, List[list], dict]]:
-    """
-    Process indicators and weights based on input parameters in the configuration.
-
-    Parameters:
-    - config: the configuration dictionary.
-    - input_matrix: the input matrix without alternatives.
-    - is_robustness_indicators: a flag indicating whether the matrix should include indicator uncertainties
-      (0 or 1).
-    - is_robustness_weights: a flag indicating whether robustness analysis is considered for the weights (0 or 1).
-    - marginal_pdf: a list of marginal probability density functions for indicators.
-    - mc_runs: number of Monte Carlo runs for robustness analysis.
[…]
