+++ mcda/models/ProMCDA.py
@@ -1,0 +1,153 @@
+import sys
+import time
+import logging
+import pandas as pd
+from typing import Tuple
+
+from mcda.configuration.configuration_validator import extract_configuration_values, check_configuration_values
+
+log = logging.getLogger(__name__)
+
+FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
+logger = logging.getLogger("ProMCDA")
+
+
+class ProMCDA:
+    def __init__(self, input_matrix:pd.DataFrame, polarity:Tuple[str], sensitivity:dict, robustness:dict,
+                 monte_carlo:dict):
+        """
+        Initialize the ProMCDA class with configuration parameters.
+
+        :param input_matrix: DataFrame containing the alternatives and criteria.
+        :param polarity: List of polarity for each indicator (+ or -).
+        :param sensitivity: Sensitivity analysis configuration.
+        :param robustness: Robustness analysis configuration.
+        :param monte_carlo: Monte Carlo sampling configuration.
+        """
+        self.input_matrix = input_matrix
+        self.polarity = polarity
+        self.sensitivity = sensitivity
+        self.robustness = robustness
+        self.monte_carlo = monte_carlo
+
+        #self.validate_input_parameters_keys # TODO: still need a formal check as made in old config class,
+                                             # maybe use some of following functions validate_
+        is_robustness_indicators = self.validate_inputs()
+        self.run_mcda(is_robustness_indicators)
+
+        self.normalized_matrix = None
+        self.aggregated_matrix = None
+        self.ranked_matrix = None
+
+        #self.logger = logging.getLogger("ProMCDA")
+
+
+    def validate_inputs(self) -> int:
+        """
+        Extract and validate input configuration parameters to ensure they are correct.
+        Return a flag indicating whether robustness analysis will be performed on indicators (1) or not (0).
+        """
+        configuration_values = extract_configuration_values(self.input_matrix, self.polarity, self.sensitivity,
+                                                            self.robustness, self.monte_carlo)
+        is_robustness_indicators = check_configuration_values(configuration_values)
+
+        # Validate input TODO: move into a different function validate_input_parameters_keys
+        #self.validate_normalization(self.sensitivity['normalization'])
+        #self.validate_aggregation(self.sensitivity['aggregation'])
+        #self.validate_robustness(self.robustness)
+
+        return is_robustness_indicators
+
+    def validate_normalization(self, f_norm):
+        """
+        Validate the normalization method.
+        """
+        valid_norm_methods = ['minmax', 'target', 'standardized', 'rank']
+        if f_norm not in valid_norm_methods:
+            raise ValueError(f"Invalid normalization method: {f_norm}. Available methods: {valid_norm_methods}")
+
+    def validate_aggregation(self, f_agg):
+        """
+        Validate the aggregation method.
+        """
+        valid_agg_methods = ['weighted_sum', 'geometric', 'harmonic', 'minimum']
+        if f_agg not in valid_agg_methods:
+            raise ValueError(f"Invalid aggregation method: {f_agg}. Available methods: {valid_agg_methods}")
+
+    def validate_robustness(self, robustness):
+        """
+        Validate robustness analysis settings.
+        """
+        if not isinstance(robustness, dict):
+            raise ValueError("Robustness settings must be a dictionary.")
+
+        # Add more specific checks based on robustness config structure
+        if robustness['on_single_weights'] == 'yes' and robustness['on_all_weights'] == 'yes':
+            raise ValueError("Conflicting settings for robustness analysis on weights.")
+
+    def normalize(self):
+        """
[…]

[…]
