### 2024-10-10 #6190 @mspada
Hi @galessiorob, I'm just checking what is happening with the paper, since I haven't seen any activity since almost a month now. Did you check the links I posted on the previous comment?
Thank you in advance for your answer and support,
Best,
Matteo
