+++ mcda/utils/utils_for_main.py
@@ -1,4 +1,3 @@
-
 import os
 import argparse
 import json
@@ -6,7 +5,7 @@ import pickle
 import random
 import logging
 import sys
-from typing import Union, Any, List, Tuple
+from typing import Union, Any, List
 from typing import Optional
 
 import numpy as np
@@ -17,8 +16,8 @@ from sklearn import preprocessing
 import mcda.utils.utils_for_parallelization as utils_for_parallelization
 import mcda.utils.utils_for_plotting as utils_for_plotting
 from mcda.configuration.config import Config
-from mcda.mcda_without_robustness import MCDAWithoutRobustness
-from mcda.mcda_with_robustness import MCDAWithRobustness
+from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
+from mcda.models.mcda_with_robustness import MCDAWithRobustness
 
 DEFAULT_INPUT_DIRECTORY_PATH = './input_files'  # present in the root directory of ProMCDA
 DEFAULT_OUTPUT_DIRECTORY_PATH = './output_files'  # present in the root directory of ProMCDA
@@ -581,10 +580,8 @@ def check_if_pdf_is_uniform(marginal_pdf: list) -> list:
     return uniform_pdf_mask
 
 
-def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name: str, index_column_values: list,
-                                           input_matrix: pd.DataFrame,
-                                           weights: Union[list, List[list], dict],
-                                           f_norm: str, f_agg: str, is_robustness_weights: int):
+def run_mcda_without_indicator_uncertainty(extracted_values: dict, is_robustness_weights: int,
+                                           weights: Union[List[str], List[pd.DataFrame], dict, None]):
     """
     Runs ProMCDA without uncertainty on the indicators, i.e. without performing
     a robustness analysis.
@@ -594,18 +591,12 @@ def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name
     and logs the completion time.
 
     Parameters:
-    - input_matrix: the input_matrix without the alternatives.
-    - index_column_name: the name of the index column of the original input matrix.
-    - index_column_values: the values of the index column of the original input matrix.
+    - extracted_values: a dictionary containing configuration values extracted from the input parameters.
+    - is_robustness_weights: a flag indicating whether robustness analysis will be performed on indicators or not.
     - weights: the normalised weights (either fixed or random sampled weights, depending on the settings).
 
-    :param input_config: dict
-    :param index_column_name: str
-    :param index_column_values: list
-    :param input_matrix: pd:DataFrame
     :param weights: Union[List[str], List[pd.DataFrame], dict, None]
-    :param f_norm: str
-    :param f_agg: str
+    :param extracted_values: dict
     :param is_robustness_weights: int
     :return: None
     """
@@ -619,11 +610,20 @@ def run_mcda_without_indicator_uncertainty(input_config: dict, index_column_name
     iterative_random_w_score_means = {}
     iterative_random_w_score_stds = {}
 
+    # Extract relevant values
+    input_matrix = extracted_values["input_matrix"]
+    alternatives_column_name = input_matrix.columns[0]
+    input_matrix = input_matrix.set_index(alternatives_column_name)
+    index_column_name = input_matrix.index.name
+    index_column_values = input_matrix.index.tolist()
+
     logger.info("Start ProMCDA without robustness of the indicators")
-    config = Config(input_config)
+    config = Config(extracted_values)
     is_sensitivity = config.sensitivity['sensitivity_on']
     is_robustness = config.robustness['robustness_on']
     mcda_no_uncert = MCDAWithoutRobustness(config, input_matrix)
+    f_norm = extracted_values["normalization"]
[…]
