+++ tests/unit_tests/test_utils_for_parallelization.py
@@ -3,7 +3,7 @@ from statistics import mean, stdev
 
 from pandas.testing import assert_frame_equal
 
-from mcda.mcda_without_robustness import *
+from mcda.models.mcda_without_robustness import *
 from mcda.utils.utils_for_parallelization import *
 
 

[…]
