+++ mcda/models/mcda_without_robustness.py
@@ -1,0 +1,157 @@
+import sys
+import copy
+import logging
+import pandas as pd
+
+from mcda.configuration.config import Config
+from mcda.mcda_functions.normalization import Normalization
+from mcda.mcda_functions.aggregation import Aggregation
+
+log = logging.getLogger(__name__)
+
+formatter = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=formatter)
+logger = logging.getLogger("ProMCDA aggregation")
+
+
+class MCDAWithoutRobustness:
+    """
+    Class MCDA without indicators' uncertainty
+
+    This class allows one to run MCDA without considering the uncertainties related to the indicators.
+    All indicators are described by the exact marginal distribution.
+    However, it's possible to have randomly sampled weights.
+    """
+
+    def __init__(self, config: dict, input_matrix: pd.DataFrame):
+        self.normalized_indicators = None
+        self.weights = None
+        self._config = copy.deepcopy(config)
+        self._input_matrix = copy.deepcopy(input_matrix)
+
+    def normalize_indicators(self, method=None) -> dict:
+        """
+        Normalize the input matrix using the specified normalization method.
+
+        Parameters:
+        - method (optional): the normalization method to use. If None, all available methods will be applied.
+          Supported methods: 'minmax', 'target', 'standardized', 'rank'.
+
+        Returns:
+        - a dictionary containing the normalized values of each indicator per normalization method.
+
+        Notes:
+        Some aggregation methods do not work with indicator values equal or smaller than zero. For that reason:
+        - for the 'minmax' method, two sets of normalized indicators are returned: one with the range (0, 1) and
+          another with the range (0.1, 1).
+        - for the 'target' method, two sets of normalized indicators are returned: one with the range (0, 1) and
+          another with the range (0.1, 1).
+        - for the 'standardized' method, two sets of normalized indicators are returned: one with the range (-inf, +inf)
+          and another with the range (0.1, +inf).
+        """
+        norm = Normalization(self._input_matrix,
+                             self._config["polarity"])
+
+        normalized_indicators = {}
+
+        if method is None or method == 'minmax':
+            indicators_scaled_minmax_01 = norm.minmax(feature_range=(0, 1))
+            # for aggregation "geometric" and "harmonic" that do not accept 0
+            indicators_scaled_minmax_without_zero = norm.minmax(feature_range=(0.1, 1))
+            normalized_indicators["minmax_without_zero"] = indicators_scaled_minmax_without_zero
+            normalized_indicators["minmax_01"] = indicators_scaled_minmax_01
+        if method is None or method == 'target':
+            indicators_scaled_target_01 = norm.target(feature_range=(0, 1))
+            indicators_scaled_target_without_zero = norm.target(
+                feature_range=(0.1, 1))  # for aggregation "geometric" and "harmonic" that do not accept 0
+            normalized_indicators["target_without_zero"] = indicators_scaled_target_without_zero
+            normalized_indicators["target_01"] = indicators_scaled_target_01
+        if method is None or method == 'standardized':
+            indicators_scaled_standardized_any = norm.standardized(
+                feature_range=('-inf', '+inf'))
+            indicators_scaled_standardized_without_zero = norm.standardized(
+                feature_range=(0.1, '+inf'))
+            normalized_indicators["standardized_any"] = indicators_scaled_standardized_any
+            normalized_indicators["standardized_without_zero"] = indicators_scaled_standardized_without_zero
+        if method is None or method == 'rank':
+            indicators_scaled_rank = norm.rank()
+            normalized_indicators["rank"] = indicators_scaled_rank
+        if method is not None and method not in ['minmax', 'target', 'standardized', 'rank']:
[…]
