+++ mcda/utils/utils_for_main.py
@@ -12,10 +12,10 @@ import numpy as np
 import pandas as pd
 from datetime import datetime
 from sklearn import preprocessing
+from sklearn.preprocessing import MinMaxScaler
 
 import mcda.utils.utils_for_parallelization as utils_for_parallelization
 import mcda.utils.utils_for_plotting as utils_for_plotting
-from mcda.configuration.config import Config
 from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.models.mcda_with_robustness import MCDAWithRobustness
 
@@ -459,12 +459,11 @@ def check_parameters_pdf(input_matrix: pd.DataFrame, config: dict, for_testing=F
     :param for_testing: bool
     :return: Union[list, None]
     """
-    config = Config(config)
 
     satisfies_condition = False
     problem_logged = False
 
-    marginal_pdf = config.monte_carlo_sampling["marginal_distribution_for_each_indicator"]
+    marginal_pdf = config["marginal_distribution_for_each_indicator"]
     is_exact_pdf_mask = check_if_pdf_is_exact(marginal_pdf)
     is_poisson_pdf_mask = check_if_pdf_is_poisson(marginal_pdf)
     is_uniform_pdf_mask = check_if_pdf_is_uniform(marginal_pdf)
@@ -616,14 +615,15 @@ def run_mcda_without_indicator_uncertainty(extracted_values: dict, is_robustness
     input_matrix = input_matrix.set_index(alternatives_column_name)
     index_column_name = input_matrix.index.name
     index_column_values = input_matrix.index.tolist()
-
-    logger.info("Start ProMCDA without robustness of the indicators")
-    config = Config(extracted_values)
-    is_sensitivity = config.sensitivity['sensitivity_on']
-    is_robustness = config.robustness['robustness_on']
-    mcda_no_uncert = MCDAWithoutRobustness(config, input_matrix)
+    input_matrix_no_alternatives = check_input_matrix(input_matrix)
+    is_sensitivity = extracted_values['sensitivity_on']
+    is_robustness = extracted_values['robustness_on']
     f_norm = extracted_values["normalization"]
     f_agg = extracted_values["aggregation"]
+
+    mcda_no_uncert\
+        = MCDAWithoutRobustness(extracted_values, input_matrix_no_alternatives)
+    logger.info("Start ProMCDA without robustness of the indicators")
 
     normalized_indicators = mcda_no_uncert.normalize_indicators() if is_sensitivity == "yes" \
         else mcda_no_uncert.normalize_indicators(f_norm)
@@ -633,15 +633,16 @@ def run_mcda_without_indicator_uncertainty(extracted_values: dict, is_robustness
             if is_sensitivity == "yes" \
             else mcda_no_uncert.aggregate_indicators(normalized_indicators, weights, f_agg)
         normalized_scores = rescale_minmax(scores)
-    elif config.robustness["on_all_weights"] == "yes" and config.robustness["robustness_on"] == "yes":
+    elif extracted_values["on_all_weights"] == "yes" and extracted_values["robustness_on"] == "yes":
         # ALL RANDOMLY SAMPLED WEIGHTS (MCDA runs num_samples times)
         all_weights_score_means, all_weights_score_stds, \
             all_weights_score_means_normalized, all_weights_score_stds_normalized = \
             _compute_scores_for_all_random_weights(normalized_indicators, is_sensitivity, weights, f_agg)
-    elif (config.robustness["on_single_weights"] == "yes") and (config.robustness["robustness_on"] == "yes"):
+    elif (extracted_values["on_single_weights"] == "yes") and (extracted_values["robustness_on"] == "yes"):
[…]

[…]
