### 2024-10-21 7742a1a
refactor: add basic uni tests and remove open socket warning
