### 2024-10-21 9d747ca
refactor: delete config.py and add a function to validate dictionary keys in config settings
