### 2024-10-22 88e31c5
refactor: add normalization as a method, failing tests
