### 2024-10-24 d9e03b1
refactor: add enums for enumerations
