+++ mcda/models/mcda_without_robustness.py
@@ -3,6 +3,8 @@ import copy
 import logging
 import pandas as pd
 
+from mcda.configuration.enums import NormalizationFunctions, OutputColumNames4Sensitivity, \
+    NormalizationNames4Sensitivity, AggregationFunctions
 from mcda.mcda_functions.normalization import Normalization
 from mcda.mcda_functions.aggregation import Aggregation
 
@@ -53,29 +55,35 @@ class MCDAWithoutRobustness:
 
         normalized_indicators = {}
 
-        if method is None or method == 'minmax':
+        if isinstance(method, NormalizationFunctions):
+            method = method.value
+        if method is None or method == NormalizationFunctions.MINMAX.value:
             indicators_scaled_minmax_01 = norm.minmax(feature_range=(0, 1))
             # for aggregation "geometric" and "harmonic" that do not accept 0
             indicators_scaled_minmax_without_zero = norm.minmax(feature_range=(0.1, 1))
-            normalized_indicators["minmax_without_zero"] = indicators_scaled_minmax_without_zero
-            normalized_indicators["minmax_01"] = indicators_scaled_minmax_01
-        if method is None or method == 'target':
+            normalized_indicators[NormalizationNames4Sensitivity.MINMAX_WITHOUT_ZERO.value] = \
+                indicators_scaled_minmax_without_zero
+            normalized_indicators[NormalizationNames4Sensitivity.MINMAX_01.value] = indicators_scaled_minmax_01
+        if method is None or method == NormalizationFunctions.TARGET.value:
             indicators_scaled_target_01 = norm.target(feature_range=(0, 1))
             indicators_scaled_target_without_zero = norm.target(
                 feature_range=(0.1, 1))  # for aggregation "geometric" and "harmonic" that do not accept 0
-            normalized_indicators["target_without_zero"] = indicators_scaled_target_without_zero
-            normalized_indicators["target_01"] = indicators_scaled_target_01
-        if method is None or method == 'standardized':
+            normalized_indicators[NormalizationNames4Sensitivity.TARGET_WITHOUT_ZERO.value] = \
+                indicators_scaled_target_without_zero
+            normalized_indicators[NormalizationNames4Sensitivity.TARGET_01.value] = indicators_scaled_target_01
+        if method is None or method == NormalizationFunctions.STANDARDIZED.value:
             indicators_scaled_standardized_any = norm.standardized(
                 feature_range=('-inf', '+inf'))
             indicators_scaled_standardized_without_zero = norm.standardized(
                 feature_range=(0.1, '+inf'))
-            normalized_indicators["standardized_any"] = indicators_scaled_standardized_any
-            normalized_indicators["standardized_without_zero"] = indicators_scaled_standardized_without_zero
-        if method is None or method == 'rank':
+            normalized_indicators[NormalizationNames4Sensitivity.STANDARDIZED_ANY.value] = indicators_scaled_standardized_any
+            normalized_indicators[NormalizationNames4Sensitivity.STANDARDIZED_WITHOUT_ZERO.value] = indicators_scaled_standardized_without_zero
+        if method is None or method == NormalizationFunctions.RANK.value:
             indicators_scaled_rank = norm.rank()
-            normalized_indicators["rank"] = indicators_scaled_rank
-        if method is not None and method not in ['minmax', 'target', 'standardized', 'rank']:
+            normalized_indicators[NormalizationNames4Sensitivity.RANK.value] = indicators_scaled_rank
+        if isinstance(method, NormalizationFunctions):
+            method = method.value
+        if method is not None and method not in [method.value for method in NormalizationFunctions]:
             logger.error('Error Message', stack_info=True)
             raise ValueError(
                 'The selected normalization method is not supported')
@@ -113,32 +121,37 @@ class MCDAWithoutRobustness:
 
[…]
