### 2024-11-01 5c0cd1d
refactor: simplify normalize and aggregate methods and draft unit tests, without robustness
