+++ mcda/models/ProMCDA.py
@@ -4,11 +4,12 @@ import logging
 import pandas as pd
 from typing import Tuple, List, Union
 
+
 from mcda.configuration.configuration_validator import extract_configuration_values, check_configuration_values, \
     check_configuration_keys
-from mcda.configuration.enums import NormalizationFunctions
-from mcda.mcda_functions.normalization import Normalization
-from mcda.utils.utils_for_main import run_mcda_without_indicator_uncertainty, run_mcda_with_indicator_uncertainty
+from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
+from mcda.utils.utils_for_main import run_mcda_without_indicator_uncertainty, run_mcda_with_indicator_uncertainty, \
+    check_input_matrix
 
 log = logging.getLogger(__name__)
 
@@ -52,8 +53,10 @@ class ProMCDA:
             print(f"Configuration Error: {e}")
             raise  # Optionally re-raise the error after logging it
 
+        self.configuration_settings = extract_configuration_values(self.input_matrix, self.polarity, self.sensitivity,
+                                                                   self.robustness, self.monte_carlo, self.output_path)
         is_robustness_indicators, is_robustness_weights, polar, weights, configuration_settings = self.validate_inputs()
-        self.run_mcda(is_robustness_indicators, is_robustness_weights, weights, configuration_settings)
+        self.run_mcda(is_robustness_indicators, is_robustness_weights, weights)
 
         self.normalized_matrix = None
         self.aggregated_matrix = None
@@ -72,80 +75,170 @@ class ProMCDA:
 
         return is_robustness_indicators, is_robustness_weights, polar, weights, configuration_values
 
-    def normalize(self, feature_range=(0, 1)) -> Union[pd.DataFrame, dict]:
+    def normalize(self, method=None) -> dict:
         """
-        Normalize the decision matrix based on the configuration `f_norm`.
+        Normalize the input data using the specified method.
 
-        If `f_norm` is a string representing a single normalization method,
-        it applies that method to the decision matrix.
-
-        If `f_norm` is a list of functions, each normalization function will be
-        applied to the input matrix sequentially, and the results will be stored
-        in a dictionary where the keys are function names.
-
-        Args:
-            feature_range (tuple): Range for normalization methods that require it, like MinMax normalization.
-                                   The range (0.1, 1) is not needed when no aggregation will follow.
+        Parameters:
+        - method (optional): The normalization method to use. If None, all available methods will be applied.
 
         Returns:
-            A single normalized DataFrame or a dictionary of DataFrames if multiple
-            normalization methods are applied.
+        - A dictionary containing the normalized values of each indicator per normalization method.
         """
-        normalization = Normalization(self.input_matrix, self.polarity)
+        input_matrix_no_alternatives = check_input_matrix(self.input_matrix)
+        mcda_without_robustness = MCDAWithoutRobustness(self.configuration_settings, input_matrix_no_alternatives)
+        normalized_values = mcda_without_robustness.normalize_indicators(method)
 
-        sensitivity_on = self.sensitivity['sensitivity_on']
-        f_norm = self.sensitivity['normalization']
-        if isinstance(f_norm, NormalizationFunctions):
-            f_norm = f_norm.value
+        return normalized_values
 
-        if sensitivity_on == "yes":
-            self.normalized_matrix = {}
-            for norm_function in f_norm:
-                self.logger.info("Applying normalization method: %s", norm_function)
-                norm_method = getattr(normalization, norm_function, None)
-                if norm_function in {NormalizationFunctions.MINMAX.value, NormalizationFunctions.STANDARDIZED.value,
-                                     NormalizationFunctions.TARGET.value}:
-                    result = norm_method(feature_range)
-                    if result is None:
[…]
