+++ tests/unit_tests/test_promcda.py
@@ -6,7 +6,8 @@ import warnings
 import pandas as pd
 
 from mcda.models.ProMCDA import ProMCDA
-from mcda.configuration.enums import NormalizationFunctions, AggregationFunctions
+from mcda.configuration.enums import NormalizationFunctions, AggregationFunctions, OutputColumnNames4Sensitivity, \
+    NormalizationNames4Sensitivity
 
 
 class TestProMCDA(unittest.TestCase):
@@ -76,45 +77,205 @@ class TestProMCDA(unittest.TestCase):
         self.assertEqual(is_robustness_indicators, 0)
         self.assertEqual(is_robustness_weights, 0)
 
-    def test_normalize_single_method(self):
-        """
-        Test normalization with a single methods.
-        Test the correctness of the output values happens in unit_tests/test_normalization.py
-        """
+    def test_normalize_all_methods(self):
         # Given
-        self.sensitivity['sensitivity_on'] = 'no'
+        self.sensitivity['sensitivity_on'] = 'yes'
+        promcda = ProMCDA(self.input_matrix, self.polarity, self.sensitivity, self.robustness, self.monte_carlo,
+                          self.output_path)
+
+        expected_keys = [method.value for method in NormalizationNames4Sensitivity]
 
         # When
-        promcda = ProMCDA(self.input_matrix, self.polarity, self.sensitivity, self.robustness, self.monte_carlo,
-                          self.output_path)
-        normalized_matrix = promcda.normalize()
+        normalized_values = promcda.normalize()
 
         # Then
-        self.assertIsInstance(normalized_matrix, pd.DataFrame)
+        self.assertCountEqual(list(set(normalized_values.keys())), expected_keys,
+                              "Not all methods were applied or extra keys found.")
+        self.assertEqual(list(normalized_values), (expected_keys))
 
-    def test_normalize_multiple_methods(self):
-        """
-        Test normalization with multiple methods.
-        Test the correctness of the output values happens in unit_tests/test_normalization.py
-        """
-
-        self.sensitivity['sensitivity_on'] = 'yes'
-        self.sensitivity['normalization'] = [method.value for method in NormalizationFunctions]
-
+    def test_normalize_specific_method(self):
+        # Given
         promcda = ProMCDA(self.input_matrix, self.polarity, self.sensitivity, self.robustness, self.monte_carlo,
-                          self.output_path)
-        normalized_matrices = promcda.normalize()
+                              self.output_path)
+        method = 'minmax'
 
-        self.assertIsInstance(normalized_matrices, dict)
-        self.assertIn(NormalizationFunctions.MINMAX.value, normalized_matrices)
-        self.assertIn(NormalizationFunctions.STANDARDIZED.value, normalized_matrices)
-        self.assertIn(NormalizationFunctions.RANK.value, normalized_matrices)
-        self.assertIn(NormalizationFunctions.TARGET.value, normalized_matrices)
+        # When
+        normalized_values = promcda.normalize(method=method)
 
-        self.assertIsInstance(normalized_matrices[NormalizationFunctions.MINMAX.value], pd.DataFrame)
-        self.assertIsInstance(normalized_matrices[NormalizationFunctions.STANDARDIZED.value], pd.DataFrame)
-        self.assertIsInstance(normalized_matrices[NormalizationFunctions.RANK.value], pd.DataFrame)
-        self.assertIsInstance(normalized_matrices[NormalizationFunctions.TARGET.value], pd.DataFrame)
[…]
