### 2024-11-06 #6190 @Flaminietta
Hi @galessiorob,

It has already been a few weeks since everything reached a standstill regarding the publication. Can we authors help in any way?

In the meantime, we are working on a new, more pytonistic version of the tool, so it would be good if we could close the first phase with the article's publication as well. We are at your disposal for any requests.

Best,
Flaminia and Matteo
