### 2024-11-12 1378354
add notebook for testing library function, change normalized_scores and scores from dict to dataframe, broken tests
