+++ mcda/models/ProMCDA.py
@@ -1,21 +1,12 @@
-import sys
 import time
-import logging
 import pandas as pd
 from typing import Tuple, List, Union
-
 
 from mcda.configuration.configuration_validator import extract_configuration_values, check_configuration_values, \
     check_configuration_keys
 from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.utils.utils_for_main import run_mcda_without_indicator_uncertainty, run_mcda_with_indicator_uncertainty, \
     check_input_matrix
-
-log = logging.getLogger(__name__)
-
-FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
-logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
-logger = logging.getLogger("ProMCDA")
 
 
 class ProMCDA:
@@ -37,7 +28,6 @@ class ProMCDA:
             df_normalized = promcda.normalize()
             df_aggregated = promcda.aggregate()
         """
-        self.logger = logging.getLogger("ProMCDA")
         self.input_matrix = input_matrix
         self.polarity = polarity
         self.sensitivity = sensitivity
@@ -75,7 +65,7 @@ class ProMCDA:
 
         return is_robustness_indicators, is_robustness_weights, polar, weights, configuration_values
 
-    def normalize(self, method=None) -> dict:
+    def normalize(self, method=None) -> pd.DataFrame:
         """
         Normalize the input data using the specified method.
 
@@ -83,15 +73,25 @@ class ProMCDA:
         - method (optional): The normalization method to use. If None, all available methods will be applied.
 
         Returns:
-        - A dictionary containing the normalized values of each indicator per normalization method.
+        - A pd.DataFrame containing the normalized values of each indicator per normalization method.
+
+        :param method: str
+        :return normalized_df: pd.DataFrame
         """
         input_matrix_no_alternatives = check_input_matrix(self.input_matrix)
         mcda_without_robustness = MCDAWithoutRobustness(self.configuration_settings, input_matrix_no_alternatives)
         normalized_values = mcda_without_robustness.normalize_indicators(method)
 
-        return normalized_values
+        if method is None:
+            normalized_df = pd.concat(normalized_values, axis=1)
+            normalized_df.columns = [f"{col}_{method}" for method, cols in normalized_values.items() for col in
+                                     input_matrix_no_alternatives.columns]
+        else:
+            normalized_df = normalized_values
 
-    def aggregate(self, normalization_method=None, aggregation_method=None, weights=None):
+        return normalized_df
+
+    def aggregate(self, normalization_method=None, aggregation_method=None, weights=None) -> pd.DataFrame:
         """
         Aggregate normalized indicators using the specified method.
 
@@ -117,126 +117,6 @@ class ProMCDA:
         return aggregated_scores
 
 
-        # def normalize(self, feature_range=(0, 1)) -> Union[pd.DataFrame, dict]:
-        #     """
-        #     Normalize the decision matrix based on the configuration `f_norm`.
-        #
-        #     If `f_norm` is a string representing a single normalization method,
-        #     it applies that method to the decision matrix.
-        #
-        #     If `f_norm` is a list of functions, each normalization function will be
-        #     applied to the input matrix sequentially, and the results will be stored
-        #     in a dictionary where the keys are function names.
-        #
-        #     Args:
-        #         feature_range (tuple): Range for normalization methods that require it, like MinMax normalization.
-        #                                The range (0.1, 1) is not needed when no aggregation will follow.
-        #
-        #     Returns:
-        #         A single normalized DataFrame or a dictionary of DataFrames if multiple
-        #         normalization methods are applied.
-        #     """
[…]
