+++ demo_in_notebook/test_import.py
@@ -1,0 +1,10 @@
+import os
+import sys
+
+# Set the package path to the parent directory
+package_path = '/Users/flaminia/Documents/work/ProMCDA'
+sys.path.append(package_path)
+
+# Now import ProMCDA from models
+from mcda.models.ProMCDA import ProMCDA
+print("Import successful!")

[…]
