### 2024-11-13 d980127
fix test_aggregate_with_specific_normalization_and_aggregation_methods
