+++ tests/unit_tests/test_promcda.py
@@ -85,7 +85,7 @@ class TestProMCDA(unittest.TestCase):
         promcda = ProMCDA(self.input_matrix, self.polarity, self.sensitivity, self.robustness, self.monte_carlo,
                           self.output_path)
 
-        # expected_keys = [method.value for method in NormalizationNames4Sensitivity]
+
         # TODO: delete if return is not a dic
         expected_suffixes = [method.value for method in NormalizationNames4Sensitivity]
 
@@ -93,9 +93,6 @@ class TestProMCDA(unittest.TestCase):
         normalized_values = promcda.normalize()
 
         # Then
-        #self.assertCountEqual(list(set(normalized_values.keys())), expected_keys,
-        #                      "Not all methods were applied or extra keys found.")
-        #self.assertEqual(list(normalized_values), (expected_keys))
         # TODO: delete if return is not a dic
         actual_suffixes = {col.split('_',1)[1] for col in normalized_values.columns}
         self.assertCountEqual(actual_suffixes, expected_suffixes,
@@ -109,26 +106,26 @@ class TestProMCDA(unittest.TestCase):
 
         # When
         normalized_values = promcda.normalize(method=method)
+        expected_keys = ['0_minmax_01', '1_minmax_01', '0_minmax_without_zero', '1_minmax_without_zero']
 
         # Then
-        expected_keys = ['Criteria 1_minmax_without_zero', 'Criteria 2_minmax_without_zero', 'Criteria 1_minmax_01',
-         'Criteria 2_minmax_01']
         self.assertCountEqual(expected_keys, list(normalized_values.keys()))
         self.assertEqual(list(normalized_values), expected_keys)
 
     def test_aggregate_all_methods(self):
         # Given
+        self.sensitivity['sensitivity_on'] = 'yes'
         promcda = ProMCDA(self.input_matrix, self.polarity, self.sensitivity, self.robustness, self.monte_carlo,
                           self.output_path)
-        aggregated_scores = promcda.aggregate(weights=self.robustness['given_weights'])
+        aggregated_scores = promcda.aggregate(normalization_method=None,
+                                              aggregation_method=None,
+                                              weights=self.robustness['given_weights'])
 
         # When
         expected_columns = [
-            'ws-minmax_01', 'ws-target_01', 'ws-standardized_any', 'ws-rank',
-            'geom-minmax_without_zero', 'geom-target_without_zero', 'geom-standardized_without_zero', 'geom-rank',
-            'harm-minmax_without_zero', 'harm-target_without_zero', 'harm-standardized_without_zero', 'harm-rank',
-            'min-standardized_any'
-        ]
+            'minmax_weighted_sum', 'target_weighted_sum', 'standardized_weighted_sum', 'rank_weighted_sum',
+            'minmax_geometric', 'minmax_minimum', 'target_geometric', 'target_minimum', 'standardized_geometric',
+            'standardized_minimum', 'rank_geometric', 'rank_minimum']
 
         # Then
         self.assertCountEqual(aggregated_scores.columns, expected_columns,
@@ -147,135 +144,14 @@ class TestProMCDA(unittest.TestCase):
                                               weights=self.robustness['given_weights'])
 
         # When
-        expected_columns = ['minmax_01_weighted_sum', 'minmax_without_zero_weighted_sum']
[…]
