### 2024-11-14 8ca4f1d
test in notebook basic methods normlize and aggregate
