### 2024-11-15 d312433
improve column names in normalize
