### 2024-11-20 7deee2d
change and simplify the __init__, add optional and default parameters
