### 2024-11-20 f7efe84
modify normalize method and test for mcda without robustness
