### 2024-11-24 #6190 @arfon
@Flaminietta – as @mspada mentioned above, it looks like you'll need to do something similar to this: https://github.com/bancaditalia/black-it/pull/21

Could you please take a look and see what you think?

/ cc @samhforbes for a paper getting stuck here
