### 2024-11-26 wetransform-os/ProMCDA#54 @mspada
Figure Input correction for JOSS Paper
Update paper-md to include JATS code for figures.
This should fix [this issue](https://github.com/openjournals/joss-reviews/issues/6190#issuecomment-2338521592).
