### 2024-11-26 68d1a0c
Update JOSS paper to avoid figure issue
