### 2024-11-26 fbf90ee
add normalization with robustness
