### 2024-11-27 93441d2
modify test for norm with robustness
