### 2024-11-27 1b83ce0
add updated md and pdf files
