### 2024-11-27 a064fc0
add and test normalization with robustness on indicators
