### 2024-11-28 91238f3
Modified caption from Figure to Table - JOSS Paper
