### 2024-11-28 a770cc1
finalize normalization
