+++ JOSS_paper/paper.md
@@ -38,23 +38,24 @@ editor_options:
 
 # Summary
 
-Multi-Criteria Decision Analysis (MCDA) is a formal process used to assist decision-makers in structuring complex decision problems and providing recommendations based on a comprehensive evaluation of alternatives. This evaluation is conducted by selecting relevant criteria and subcriteria, which are then aggregated according to the preferences of the decision-makers to produce a ranking or classification of the alternatives (@roy_decision_1996; @bouyssou_problem_2006). A wide range of MCDA methods are available in the literature for integrating information to classify alternatives into preference classes or rank them from best to worst (@cinelli_proper_2022). Among these, composite indicators (CIs) are commonly used synthetic measures for ranking and benchmarking alternatives across complex concepts (@greco_methodological_2019). Examples of CI applications include environmental quality assessment (@otoiu_proposing_2018), resilience of energy supply (@gasser_comprehensive_2020), sustainability (@volkart_interdisciplinary_2016), and global competitiveness (@klaus_schwab_global_2018).
+Multi-Criteria Decision Analysis (MCDA) is a formal process used to assist decision-makers in structuring complex decision problems and providing recommendations based on a comprehensive evaluation of alternatives. This evaluation is conducted by selecting relevant criteria and subcriteria, which are then aggregated according to the preferences of the decision-makers to produce a ranking or classification of the alternatives [@roy_decision_1996; @bouyssou_problem_2006]. A wide range of MCDA methods are available in the literature for integrating information to classify alternatives into preference classes or rank them from best to worst [@cinelli_proper_2022]. Among these, composite indicators (CIs) are commonly used synthetic measures for ranking and benchmarking alternatives across complex concepts [@greco_methodological_2019]. Examples of CI applications include environmental quality assessment [@otoiu_proposing_2018], resilience of energy supply [@gasser_comprehensive_2020], sustainability [@volkart_interdisciplinary_2016], and global competitiveness [@klaus_schwab_global_2018].
 
-However, the final ranking of alternatives in MCDA can be influenced by various factors such as uncertainty in the criteria, the choice of weights assigned to them, and the selection of methods for normalization and aggregation to construct CIs (@cinelli_mcda_2020). To address these challenges, the `ProMCDA` Python module has been developed to allow decision-makers to explore the sensitivity and robustness of CI results in a user-friendly manner. This tool facilitates sensitivity analysis related to the choice of normalization and aggregation methods and accounts for uncertainty in criteria and weights, providing a systematic approach to understanding the impact of these factors on decision outcomes.
+However, the final ranking of alternatives in MCDA can be influenced by various factors such as uncertainty in the criteria, the choice of weights assigned to them, and the selection of methods for normalization and aggregation to construct CIs [@langhans_method_2014; @cinelli_mcda_2020]. To address these challenges, the `ProMCDA` Python module has been developed to allow decision-makers to explore the sensitivity and robustness of CI results in a user-friendly manner. This tool facilitates sensitivity analysis related to the choice of normalization and aggregation methods and accounts for uncertainty in criteria and weights, providing a systematic approach to understanding the impact of these factors on decision outcomes.
 
 # Statement of need
 
[…]
