### 2024-11-28 #6190 @Flaminietta
Dear @arfon,

the last PR merged into `main` should fix the problems.
You can see it [here](https://github.com/wetransform-os/ProMCDA/pull/55).

Let us know if there is any other issue or if the publication can proceed now.

Best regards,
Flaminia and Matteo
