### 2024-12-03 d587fe3
add aggregation with specific method
