+++ mcda/models/ProMCDA.py
@@ -5,7 +5,7 @@ from typing import Tuple, List, Union, Optional
 from build.lib.mcda.mcda_with_robustness import MCDAWithRobustness
 from mcda.configuration.configuration_validator import extract_configuration_values, check_configuration_values, \
     check_configuration_keys
-from mcda.configuration.enums import PDFType, NormalizationFunctions
+from mcda.configuration.enums import PDFType, NormalizationFunctions, AggregationFunctions
 from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.utils import utils_for_parallelization
 from mcda.utils.utils_for_main import run_mcda_without_indicator_uncertainty, run_mcda_with_indicator_uncertainty, \
@@ -75,6 +75,8 @@ class ProMCDA:
         self.normalized_values_with_robustness = None
         self.scores = None
 
+        self.input_matrix_no_alternatives = check_input_matrix(self.input_matrix)
+
     # def validate_inputs(self) -> Tuple[int, int, list, Union[list, List[list], dict], dict]:
     #     """
     #     Extract and validate input configuration parameters to ensure they are correct.
@@ -102,26 +104,23 @@ class ProMCDA:
         Returns:
         - A pd.DataFrame containing the normalized values of each indicator per normalization method,
           if no robustness on indicators is performed.
-        - A dictionary containing the normalized values of each indicator per normalization method,
-          if robustness on indicators is performed.
 
         :param method: NormalizationFunctions
-        :return normalized_df: pd.DataFrame
+        :return normalized_df: pd.DataFrame or string
         """
-        input_matrix_no_alternatives = check_input_matrix(self.input_matrix)
 
         if not self.robustness_indicators:
-            mcda_without_robustness = MCDAWithoutRobustness(self.polarity, input_matrix_no_alternatives)
+            mcda_without_robustness = MCDAWithoutRobustness(self.polarity, self.input_matrix_no_alternatives)
             self.normalized_values_without_robustness = mcda_without_robustness.normalize_indicators(method)
 
             return self.normalized_values_without_robustness
 
-        elif self.robustness_indicators is not None and self.robustness_weights is None:
-            check_parameters_pdf(input_matrix_no_alternatives, self.marginal_distributions, for_testing=False)
+        elif self.robustness_indicators and not self.robustness_weights:
+            check_parameters_pdf(self.input_matrix_no_alternatives, self.marginal_distributions, for_testing=False)
             is_exact_pdf_mask = check_if_pdf_is_exact(self.marginal_distributions)
             is_poisson_pdf_mask = check_if_pdf_is_poisson(self.marginal_distributions)
 
-            mcda_with_robustness = MCDAWithRobustness(input_matrix_no_alternatives, self.marginal_distributions,
+            mcda_with_robustness = MCDAWithRobustness(self.input_matrix_no_alternatives, self.marginal_distributions,
                                                       self.num_runs, is_exact_pdf_mask, is_poisson_pdf_mask,
                                                       self.random_seed)
             n_random_input_matrices = mcda_with_robustness.create_n_randomly_sampled_matrices()
@@ -141,7 +140,6 @@ class ProMCDA:
             raise ValueError(
                 "Inconsistent configuration: 'robustness_weights' and 'robustness_indicators' are both enabled.")
 
-
     def get_normalized_values_with_robustness(self) -> Optional[pd.DataFrame]:
         """
         Getter method to access normalized values when robustness on indicators is performed.
@@ -151,29 +149,57 @@ class ProMCDA:
         """
         return getattr(self, 'normalized_values_with_robustness', None)
 
-
-    def aggregate(self, aggregation_method=None, weights=None) -> pd.DataFrame:
[…]
