### 2024-12-04 d95aaa9
add aggregation without a  specific method
