### 2024-12-11 #6190 @Flaminietta
Dear @arfon and @samhforbes, any news about the publication after we fixed the last issue? We hope to see it soon 🥳
