### 2024-12-11 #6190 @mspada
Hello @samhforbes,

we corrected the error that happened 1 hour ago.
For some reasons the figure files where not anymore in the folder.
We added them back and now it should work out with the publication (I hope so :-))

Best,
Matteo and Flaminia
