+++ mcda/models/ProMCDA.py
@@ -1,20 +1,27 @@
+import logging
+import sys
 import time
 import pandas as pd
 from typing import Tuple, List, Union, Optional
 
 from build.lib.mcda.mcda_with_robustness import MCDAWithRobustness
 from mcda.configuration.configuration_validator import extract_configuration_values, check_configuration_values, \
-    check_configuration_keys
+    check_configuration_keys, check_indicator_weights_polarities, process_indicators_and_weights
 from mcda.configuration.enums import PDFType, NormalizationFunctions, AggregationFunctions
+from mcda.models import mcda_with_robustness
 from mcda.models.mcda_without_robustness import MCDAWithoutRobustness
 from mcda.utils import utils_for_parallelization
 from mcda.utils.utils_for_main import run_mcda_without_indicator_uncertainty, run_mcda_with_indicator_uncertainty, \
-    check_input_matrix, check_if_pdf_is_exact, check_if_pdf_is_poisson, check_parameters_pdf
+    check_input_matrix, check_if_pdf_is_exact, check_if_pdf_is_poisson, check_parameters_pdf, rescale_minmax
 
+log = logging.getLogger(__name__)
+formatter = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
+logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=formatter)
+logger = logging.getLogger("ProMCDA")
 
 class ProMCDA:
     def __init__(self, input_matrix: pd.DataFrame, polarity: Tuple[str, ...], robustness_weights: Optional[bool] = None,
-                 robustness_indicators: Optional[bool] = None,
+                 robustness_single_weights: Optional[bool] = None, robustness_indicators: Optional[bool] = None,
                  marginal_distributions: Optional[Tuple[PDFType, ...]] = None,
                  num_runs: Optional[int] = 10000, num_cores: Optional[int] = 1, random_seed: Optional[int] = 43):
         """
@@ -27,6 +34,8 @@ class ProMCDA:
         # Optional parameters
         :param robustness_weights: Boolean flag indicating whether to perform robustness analysis on weights
                                    (True or False).
+        :param robustness_single_weights: Boolean flag indicating whether to perform robustness analysis on one single
+                                   weight at time (True or False).
         :param robustness_indicators: Boolean flag indicating whether to perform robustness analysis on indicators
                                       (True or False).
         :param marginal_distributions: Tuple of marginal distributions, which describe the indicators
@@ -66,6 +75,7 @@ class ProMCDA:
         self.input_matrix = input_matrix
         self.polarity = polarity
         self.robustness_weights = robustness_weights
+        self.robustness_single_weights = robustness_single_weights
         self.robustness_indicators = robustness_indicators
         self.num_runs = num_runs
         self.marginal_distributions = marginal_distributions
@@ -73,7 +83,11 @@ class ProMCDA:
         self.random_seed = random_seed
         self.normalized_values_without_robustness = None
         self.normalized_values_with_robustness = None
-        self.scores = None
+        self.aggregated_scores = None
+        self.all_indicators_scores_means = None
+        self.all_indicators_scores_stds = None
+        self.all_indicators_means_scores_normalized = None
+        self.all_indicators_scores_stds_normalized = None
 
         self.input_matrix_no_alternatives = check_input_matrix(self.input_matrix)
 
@@ -132,7 +146,7 @@ class ProMCDA:
                 n_normalized_input_matrices = utils_for_parallelization.parallelize_normalization(
                     n_random_input_matrices, self.polarity, normalization_method)
 
-                self.normalized_values_with_robustness = n_normalized_input_matrices
[…]
