+++ requirements.txt
@@ -1,13 +1,16 @@
 # Requirements automatically generated by pigar.
 # https://github.com/damnever/pigar.
 
-numpy
-pandas
-plotly
-pytest
-scikit-learn
-scipy
-Pillow
+numpy~=2.0.2
+pandas~=2.2.3
+plotly~=5.24.1
+pytest~=8.3.3
+scikit-learn~=1.5.2
+scipy~=1.13.1
+Pillow~=11.0.0
 kaleido
 
 
+
+matplotlib~=3.9.3
+setuptools~=68.2.0

[…]
