### 2024-12-12 39c9061
fix unit test for aggregation, all variations of it
