+++ tests/unit_tests/test_promcda.py
@@ -225,15 +225,16 @@ class TestProMCDA(unittest.TestCase):
             random_seed=self.random_seed
         )
         promcda.normalize(normalization_method)
-        aggregated_scores = promcda.aggregate(aggregation_method=aggregation_method)
-        expected_columns = ['minmax_weighted_sum']
+        promcda.aggregate(aggregation_method=aggregation_method)
+        aggregated_scores, aggregated_stds = promcda.get_aggregated_values_with_robustness()
+        expected_columns = ['ws-minmax_01']
 
         # Then
         self.assertCountEqual(aggregated_scores.columns, expected_columns,
                                   "Only specified methods should be applied.")
         self.assertTrue(
-            (aggregated_scores['minmax_weighted_sum'] >= 0).all() and (
-             aggregated_scores['minmax_weighted_sum'] <= 1).all(),
+            (aggregated_scores['ws-minmax_01'] >= 0).all() and (
+             aggregated_scores['ws-minmax_01'] <= 1).all(),
                 "Values should be in the range [0, 1] for minmax normalization with weighted sum.")
 
 
