### 2024-12-18 #6190 @mspada
Hi @galessiorob, @arfon, @samhforbes,

It seems to be working now!
However, the layout could use some adjustment—specifically, Tables 1 and 2 should ideally be on the same page.
What’s the best way to address this?
Should we make the corrections ourselves, or is there editorial support available for this?

Best,
Matteo and Flaminia
