+++ mcda/mcda_run.py
@@ -1,169 +1,0 @@
-#! /usr/bin/env python3
-
-"""
-This script serves as the main entry point for running all pieces of functionality in a consequential way by
-following the settings given in the configuration file 'configuration.json'.
-
-Usage (from root directory):
-    $ python3 -m mcda.mcda_run -c configuration.json
-"""
-
-import time
-import logging
-
-from mcda.configuration.config import Config
-from mcda.utils.utils_for_main import *
-from mcda.utils.utils_for_plotting import *
-from mcda.utils.utils_for_parallelization import *
-
-log = logging.getLogger(__name__)
-
-FORMATTER: str = '%(levelname)s: %(asctime)s - %(name)s - %(message)s'
-logging.basicConfig(stream=sys.stdout, level=logging.DEBUG, format=FORMATTER)
-logger = logging.getLogger("ProMCDA")
-
-# noinspection PyTypeChecker
-def main(input_config: dict):
-    """
-    Execute the ProMCDA (Probabilistic Multi-Criteria Decision Analysis) process.
-
-    Parameters:
-    - input_config : Configuration parameters for the ProMCDA process.
-
-    Raises:
-    - ValueError: If there are issues with the input matrix, weights, or indicators.
-
-    This function performs the ProMCDA process based on the provided configuration.
-    It handles various aspects such as the sensitivity analysis and the robustness analysis.
-    The results are saved in output files, and plots are generated to visualize the scores and rankings.
-
-    Note: Ensure that the input matrix, weights, polarities and indicators (with or without uncertainty)
-    are correctly specified in the input configuration.
-
-    :param input_config: dict
-    :return: None
-    """
-    is_sensitivity = None
-    is_robustness = None
-    is_robustness_indicators = 0
-    is_robustness_weights = 0
-    f_norm = None
-    f_agg = None
-    marginal_pdf = []
-
-    t = time.time()
-
-    # Extracting relevant configuration values
-    config = Config(input_config)
-    input_matrix = read_matrix(config.input_matrix_path)
-    index_column_name = input_matrix.index.name
-    index_column_values = input_matrix.index.tolist()
-    polar = config.polarity_for_each_indicator
-    is_sensitivity = config.sensitivity['sensitivity_on']
-    is_robustness = config.robustness['robustness_on']
-    mc_runs = config.monte_carlo_sampling["monte_carlo_runs"]
-    random_seed = config.monte_carlo_sampling["random_seed"]
-
-    # Check for sensitivity-related configuration errors
-    if is_sensitivity == "no":
-        f_norm = config.sensitivity['normalization']
-        f_agg = config.sensitivity['aggregation']
-        check_config_error(f_norm not in ['minmax', 'target', 'standardized', 'rank'],
-                           'The available normalization functions are: minmax, target, standardized, rank.')
-        check_config_error(f_agg not in ['weighted_sum', 'geometric', 'harmonic', 'minimum'],
-                           'The available aggregation functions are: weighted_sum, geometric, harmonic, minimum.'
-                           '\nWatch the correct spelling in the configuration file.')
-        logger.info("ProMCDA will only use one pair of norm/agg functions: " + f_norm + '/' + f_agg)
-    else:
-        logger.info("ProMCDA will use a set of different pairs of norm/agg functions")
-
-    # Check for robustness-related configuration errors
-    if is_robustness == "no":
-        logger.info("ProMCDA will run without uncertainty on the indicators or weights")
-        logger.info("Read input matrix without uncertainties at {}".format(config.input_matrix_path))
-    else:
-        check_config_error((config.robustness["on_single_weights"] == "no" and
-                            config.robustness["on_all_weights"] == "no" and
-                            config.robustness["on_indicators"] == "no"),
[…]
