+++ mcda/utils/utils_for_parallelization.py
@@ -134,21 +134,37 @@ def normalize_indicators_in_parallel(norm: object, method=None) -> dict:
     indicators_scaled_target_without_zero = None
     indicators_scaled_rank = None
 
+    def _rename_columns(df, method_name):
+        """ Helper function to rename columns based on the normalization method """
+        if df is not None:
+            df.columns = [f"{col}_{method_name}" for col in df.columns.tolist()]
+        return df
+
     if method is None or method == NormalizationFunctions.MINMAX:
         indicators_scaled_minmax_01 = norm.minmax(feature_range=(0, 1))
+        indicators_scaled_minmax_01 = _rename_columns(indicators_scaled_minmax_01, "minmax_01")
         # for aggregation "geometric" and "harmonic" that accept no 0
         indicators_scaled_minmax_without_zero = norm.minmax(feature_range=(0.1, 1))
+        indicators_scaled_minmax_without_zero = _rename_columns(indicators_scaled_minmax_without_zero,
+                                                               "minmax_without_zero")
     if method is None or method == NormalizationFunctions.TARGET:
         indicators_scaled_target_01 = norm.target(feature_range=(0, 1))
+        indicators_scaled_target_01 = _rename_columns(indicators_scaled_target_01, "target_01")
         # for aggregation "geometric" and "harmonic" that accept no 0
         indicators_scaled_target_without_zero = norm.target(feature_range=(0.1, 1))
+        indicators_scaled_target_without_zero = _rename_columns(indicators_scaled_target_without_zero,
+                                                               "target_without_zero")
     if method is None or method == NormalizationFunctions.STANDARDIZED:
         indicators_scaled_standardized_any = norm.standardized(
             feature_range=('-inf', '+inf'))
+        indicators_scaled_standardized_any = _rename_columns(indicators_scaled_standardized_any, "standardized_any")
         indicators_scaled_standardized_without_zero = norm.standardized(
             feature_range=(0.1, '+inf'))
+        indicators_scaled_standardized_without_zero = _rename_columns(indicators_scaled_standardized_without_zero,
+                                                                     "standardized_without_zero")
     if method is None or method == NormalizationFunctions.RANK:
         indicators_scaled_rank = norm.rank()
+        indicators_scaled_rank = _rename_columns(indicators_scaled_rank, "rank")
     if method is not None and method not in [e for e in NormalizationFunctions]:
         logger.error('Error Message', stack_info=True)
         raise ValueError('The selected normalization method is not supported')
@@ -168,7 +184,8 @@ def normalize_indicators_in_parallel(norm: object, method=None) -> dict:
     return normalized_indicators
 
 
-def aggregate_indicators_in_parallel(agg: object, normalized_indicators: dict, method: Optional[AggregationFunctions] = None) -> pd.DataFrame:
+def aggregate_indicators_in_parallel(agg: object, normalized_indicators: dict,
+                                     method: Optional[AggregationFunctions] = None) -> pd.DataFrame:
     """
     Aggregate normalized indicators in parallel using different aggregation methods.
 
@@ -208,27 +225,63 @@ def aggregate_indicators_in_parallel(agg: object, normalized_indicators: dict, m
                  'geom-minmax_without_zero', 'geom-target_without_zero', 'geom-standardized_without_zero', 'geom-rank',
                  'harm-minmax_without_zero', 'harm-target_without_zero', 'harm-standardized_without_zero', 'harm-rank',
                  'min-standardized_any']  # same order as in the following loop
-    for key, values in normalized_indicators.items():
+
+    if isinstance(normalized_indicators, dict): # robustness on indicators
+        for key, values in normalized_indicators.items():
+            if method is None or method == AggregationFunctions.WEIGHTED_SUM:
+                # ws goes only with some specific normalizations
[…]

[…]
