### 2024-12-18 #6190 @samhforbes
Hi @mspada
Yes sorry I could see it worked - but yes, we don't really have person-power for improving the look of the articles. If you need them to appear together you can make that change, although I can see that might be a challenge given the size of each of the tables? Just let us know how you'd like to proceed
