### 2024-12-19 17596e6
fix aggregation with robustness on indicators and weights
