### 2024-12-19 eed0195
remve hard coded column names and use enums
