### 2024-12-20 #6190 @mspada
Hi @samhforbes,
I'll try to give the paper a nicer outlook today.
Just a question: can I use the editorialbot to generate the pdf myself to check how it will looks like?
Thanks
