### 2024-12-20 f6b2fe6
JOSS Paper - Editing Layout by moving Table 2
