### 2024-12-20 #6190 @samhforbes
yes you can with the generate pdf command!
