### 2024-12-20 #6190 @samhforbes
@mspada looks like an issue with the table metadata. Tagging @arfon in case he knows more about this!
