### 2024-12-20 #6190 @mspada
@samhforbes,
Thanks, I saw the issue, but I do not understand why it is happening, since the IDs are defined only once in the .md file.
