### 2024-12-20 #6190 @mspada
@samhforbes, @arfon,
I removed the IDs and the editorialbot is able to generate the pdf correctly.
Could you please check if now the acceptance work?
Thank you!
