### 2025-01-14 #6190 @mspada
Dear @samhforbes and @arfon,
first of all, Happy New Year!
We are just checking the status of the paper, since we corrected the IDs issue and it should work out now (we hope).
Could you please check?

Thanks and Best,
Matteo and Flaminia
