### 2025-01-14 #6190 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
