### 2025-01-19 #6190 @Flaminietta
Dear all,

I can only echo what Matteo said, thank you for your work!

Best,
Flaminia

> On 19. Jan 2025, at 17:33, Matteo Spada ***@***.***> wrote:
>
>
> Thank you for the acceptance and the work and support @samhforbes and @galessiorob!
> I'd like also to thank you @limengbinggz @B3J4y and @paulrougieux for the reviews, which help us improving both the manuscript and the code!
